### 2024-12-07 #220 @canorve
GALFITools: a library for GALFIT
Submitting Author: (@canorve)
Package Name: GALFITools
One-Line Description of Package: A library for efficient data processing customized for the GALFIT package
Repository Link:  https://github.com/canorve/GALFITools
Version submitted:  v1.15.0
EiC: @SimonMolinsky, @coatless
Reviewer 2: @ tepickering
Archive: https://zenodo.org/records/17010007
JOSS DOI: 10.21105/joss.09534
Version accepted: 1.18.0

---

[GALFIT](https://users.obs.carnegiescience.edu/peng/work/galfit/galfit.html), a well-established two-dimensional image fitting algorithm (Peng et al. 2002, AJ, 124, 266), is integral to precise modeling of galaxy surface brightness in astronomical images. To optimize GALFIT's utility, GALFITools provides a suite of Python routines that streamline input and output parsing for enhanced efficiency and usability.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

GALFITools enhances functionality with a range of features, including mask creation, PSF generation, initial parameter estimation, galaxy image model visualizaton, multigaussian expansion (MGE) fitting, and calculation of sky background along with other key photometric parameters.

This a tool for astronomers to streamline image data processing and enhance interpretation of GALFIT outputs

No, similar tools act as GALFIT wrappers but focus on different objectives, such as automating GALFIT for large galaxy samples.

https://github.com/pyOpenSci/software-submission/issues/216

https://zenodo.org/records/13994492
