### 2024-12-09 #220 @SimonMolinsky
---
---

@canorve I've checked your package and have two issues that should be addressed before we assign the editor and reviewers:

1. You have provided documentation, but I see only raw `rst` files. Have you built docs somewhere? Do you have a link to the build docs? Here is the guide that might help you: https://www.pyopensci.org/python-package-guide/documentation/hosting-tools/intro.html
2. Are your tests automated? Have you considered using GitHub Actions for this? Here's an example of how to set up automated tests: https://www.pyopensci.org/python-package-guide/tests/tests-ci.html

We will go forward with the review when you address those two points!
