### 2025-01-18 148a3df
removed bug related to the incompatibility of number of components in constraint file
 Changes to be committed:
	modified:   MakePSF.py
	modified:   ../mge/mge2galfit.py
