+++ .github/workflows/run-test.yml
@@ -1,0 +1,45 @@
+name: Pytest unit/integration
+
+on:
+  pull_request:
+  push:
+    branches:
+      - main
+
+# Use bash by default in all jobs
+defaults:
+  run:
+    shell: bash
+
+jobs:
+  build-test:
+    name: Test Run (${{ matrix.python-version }}, ${{ matrix.os }})
+    runs-on: ${{ matrix.os }}
+    strategy:
+      fail-fast: false
+      matrix:
+        os: ["ubuntu-latest", "macos-latest", "windows-latest"]
+        python-version: ["3.9", "3.10", "3.11"]
+
+    steps:
+      - uses: actions/checkout@v4
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v4
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          python -m pip install tox
+          #python -m pip install nox
+      - name: List installed packages
+        run: pip list
+      - name: Run tests with Tox 
+        run: |
+          tox -e default
+          #    nox -s tests-${{ matrix.python-version }}
+      # You only need to upload code coverage once to codecov unless you have a
+      # more complex build that you need coverage for.
+      #  - name: Upload coverage to Codecov
+      #  if: ${{ matrix.os == 'ubuntu-latest' &&  matrix.python-version == '3.10'}}
+      #  uses: codecov/codecov-action@v3
