### 2025-01-24 e6e38aa
removed unnecesary file  Changes to be committed: 	deleted:    splderiv.py
