### 2025-01-27 e462219
reformatted for docstring  Changes to be committed: 	modified:   src/galfitools/galin/galfit.py 	modified:   src/galfitools/galin/initgal.py 	modified:   src/galfitools/galout/getRads.py 	modified:   src/galfitools/mge/mge2galfit.py
