### 2025-01-28 b9f51ec
docstring format updated
 On branch master
 Changes to be committed:
	modified:   api.rst
	modified:   howto.rst
	modified:   index.rst
