### 2025-01-30 f8bd3af
plot default removed for mge_fit  On branch master  Changes to be committed: 	modified:   mge2galfit.py
