### 2025-02-03 #220 @SimonMolinsky
Hi @canorve

How are you? Could you tell me how updating documentation and automated tests is going?
