### 2025-02-03 #220 @canorve
Hello @SimonMolinsky,

The documentation and GitHub Actions are new to me, but I'm working on them. The GitHub Actions workflow is already implemented, though I'm not sure if it's working properly. Currently, I'm following the Read the Docs documentation.
