### 2025-02-15 e64b725
readme updated
 Changes to be committed:
	modified:   README.rst
 modified:   docs/usage.rst
