+++ README.rst
@@ -237,15 +237,6 @@ functions within your own scripts can be found here:
 
 
 
-
------------------
-**Contribute**
-~~~~~~~~~~~~~~~~~
-
-
-Issue tracker: `https://github.com/canorve/GALFITools/issues <https://github.com/canorve/GALFITools/issues>`__. 
-source code: `https://github.com/canorve/GALFITools <https://github.com/canorve/GALFITools>`__. 
-
 -----------------
 
 **Community**
