+++ README.rst
@@ -179,6 +179,7 @@ model consisting of three Sersic components:
 **USAGE GUIDE**
 ~~~~~~~~~~~~~~~~~
 
+For full details read the documentation `https://galfitools.readthedocs.io/en/latest/ <https://galfitools.readthedocs.io/en/latest/>`__.
 
 For comprehensive insights into GALFITools' repertoire 
 of routines and their optimal deployment, it is 
@@ -234,6 +235,7 @@ The detailed instructions for utilizing these
 functions within your own scripts can be found here: 
 
 `API <docs/api.rst>`__
+
 
 
 
