+++ README.rst
@@ -245,9 +245,9 @@ functions within your own scripts can be found here:
 ~~~~~~~~~~~~~~~~~
 
 
-Issue tracker: `https://github.com/canorve/GALFITools/issues <https://github.com/canorve/GALFITools/issues>`__. 
+- Issue tracker: `https://github.com/canorve/GALFITools/issues <https://github.com/canorve/GALFITools/issues>`__. 
 
-source code: `https://github.com/canorve/GALFITools <https://github.com/canorve/GALFITools>`__. 
+- Source code: `https://github.com/canorve/GALFITools <https://github.com/canorve/GALFITools>`__. 
 
 
 
