+++ docs/usage.rst
@@ -5,7 +5,6 @@
 ============================
 
 
-
 .. contents::
    :depth: 3
 ..
