+++ docs/index.rst
@@ -47,7 +47,8 @@ Indices and tables
 * :ref:`api`
 
 
-.. __GALFIT: https://users.obs.carnegiescience.edu/peng/work/galfit/galfit.html
+
+.. _GALFIT: https://users.obs.carnegiescience.edu/peng/work/galfit/galfit.html
 
 .. _toctree: https://www.sphinx-doc.org/en/master/usage/restructuredtext/directives.html
 .. _reStructuredText: https://www.sphinx-doc.org/en/master/usage/restructuredtext/basics.html
