+++ README.rst
@@ -35,11 +35,9 @@
    :target: CODE_OF_CONDUCT.md 
 
 
-==========
-GALFITools
-==========
-
-
+============
+GALFITOOLS
+============
 
 `GALFIT`_, a well-established two-dimensional image fitting algorithm, 
 as outlined in the work by Peng et al. (2002, AJ, 124, 266), 
@@ -49,10 +47,6 @@ as an collection of Python routines. These routines
 enhances the input and output parsing associated with GALFIT.
 
 
-
-.. _GALFIT: https://users.obs.carnegiescience.edu/peng/work/galfit/galfit.html
-
-
 GALFITools extends its utility through an array of functionalities, 
 including the facilitation of mask creation, star selection for PSFs, generation 
 of multiple initial parameters, simulate galaxy images, multigaussian 
@@ -60,50 +54,22 @@ expansion (MGE) fitting, as well as computation of sky background
 and other pertinent photometric variables.
 
 
-
-
---------------------------------
-
-**Installation Instructions**
--------------------------------
+.. _GALFIT: https://users.obs.carnegiescience.edu/peng/work/galfit/galfit.html
 
 
 
-The python libraries required are:
+-----------------------------------
 
--  numpy
--  astropy
--  scipy
--  matplotlib
--  mgefit
+**INSTALLATION INSTRUCTIONS**
+-----------------------------------
 
+First of all, install `GALFIT`_ if you haven't done so. Check
+instructions `here <https://users.obs.carnegiescience.edu/peng/work/galfit/galfit.html`__.
+Make sure that GALFIT can run in any path in your *linux/macOS* terminal.
 
+*Optionally*, consider setting up a virtual environment.
 
-Install `GALFIT`_ if you haven't done so. Make sure
-that GALFIT can run in any path in your terminal.
-
-
-*Optionally*, the establishment of a virtual environment can be considered.
-
-
-The latest release of GALFITools can be downloaded 
-and subsequently installed via one of the following methods:
-
-
-::
-
-   cd GALFITools 
-   pip install . 
-
-or
-
-::
-
-   cd GALFITools 
-   python setup.py install
-
-
-Alternatively, you can install it via pip:
+Install GALFITools via pip:
 
 
 ::
@@ -111,8 +77,7 @@ Alternatively, you can install it via pip:
    pip install GALFITools 
 
 
-
-Here is tutorial for the installation of GALFITools:
+Here is video tutorial for the installation of GALFITools:
 
 
 .. image:: https://img.youtube.com/vi/rqZLxR1yRCs/maxresdefault.jpg
@@ -121,35 +86,15 @@ Here is tutorial for the installation of GALFITools:
 
 
 
+--------------------------------
 
+**GET STARTED WITH GALFITOOLS**
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-In conjunction with the installation, a compilation of pertinent 
-shell `commands <docs/usage.rst>`__ will be incorporated. Subsequently, a comprehensive 
-evaluation of GALFITools' performance can be conducted through 
-automated tests using the following procedure:
-
-To run the tests locally, install and invoke tox:
-
-::
-   
-   pip install tox
-
-
-run the tests:
-
-::
-
-    tox 
-
------------------
-
-**Get Started with GALFITools**
-~~~~~~~~~~~~~~~~~
-
-Once installed, you can test galfitools using any of 
+Once installed, you can test GALFITools using any of 
[…]
