+++ README.rst
@@ -86,7 +86,7 @@ Here is video tutorial for the installation of GALFITools:
 
 
 
---------------------------------
+------------------------------------
 
 **GET STARTED WITH GALFITOOLS**
 -----------------------------------
@@ -117,19 +117,19 @@ model consisting of three Sersic components:
   The radius at 50% of light is 199.64 pixels or 49.91 " 
 
 
------------------
+-------------------
 
 **DOCUMENTATION**
------------------------------------
+-------------------
 
-full documentation along usage guide can be found in `https://galfitools.readthedocs.io/en/latest/ <https://galfitools.readthedocs.io/en/latest/>`__.
+Full documentation along usage guide can be found in `https://galfitools.readthedocs.io/en/latest/ <https://galfitools.readthedocs.io/en/latest/>`__.
 
 
 
 ------------------------------
 
 **CONTRIBUTING AND SUPPORT**
------------------------------------
+-------------------------------
 
 Let us know if you have issues, questions or innovative suggestions arise, 
 we encourage you to reach out via email to  canorve [at] gmail [dot] com  
@@ -190,8 +190,8 @@ EllipSect tool:
 
 ====
 
-Note
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+NOTE
+~~~~~~
 
 This project has been set up using PyScaffold 4.2.1. For details and usage
 information on PyScaffold see https://pyscaffold.org/.
