+++ README.rst
@@ -144,10 +144,9 @@ GALFITools, for the betterment of the *GALFIT* community and beyond.
 
 - Source code: `https://github.com/canorve/GALFITools <https://github.com/canorve/GALFITools>`__. 
 
-- :doc: `Contributing <contributing>`
+:doc: `Contributing <contributing>`
 
 - :doc: `Code of conduct <../CODE_OF_CONDUCT.md>`
-
 
 
 --------------
