+++ README.rst
@@ -145,7 +145,7 @@ GALFITools, for the betterment of the *GALFIT* community and beyond.
 - Source code: `https://github.com/canorve/GALFITools <https://github.com/canorve/GALFITools>`__. 
 
 :doc: `Contributing <contributing>`
-
+See :doc:`contributing` 
 - :doc: `Code of conduct <../CODE_OF_CONDUCT.md>`
 
 
