+++ README.rst
@@ -144,11 +144,8 @@ GALFITools, for the betterment of the *GALFIT* community and beyond.
 
 - Source code: `https://github.com/canorve/GALFITools <https://github.com/canorve/GALFITools>`__. 
 
-:doc: `Contributing <contributing>`
-See :doc:`contributing` 
+:doc: `Contributing <contributing>`  
 
-
-   Contributions & Help <contributing>
 
 --------------
 
