+++ README.rst
@@ -146,7 +146,6 @@ GALFITools, for the betterment of the *GALFIT* community and beyond.
 
 - `Contributions and help <https://galfitools.readthedocs.io/en/latest/contributing.html>`__. 
 
-- `contributing <CONTRIBUTING>`__ 
 
 --------------
 
