+++ CHANGELOG.rst
@@ -2,14 +2,6 @@
 Changelog
 =========
 
-Version 0.1
-===========
-
-- Feature A added
-- FIX: nasty bug #1729 fixed
-- add your changes here!
-
-
 
 Version 0.15.2
 ===============
@@ -28,6 +20,87 @@ modeling. With GALFITools, you can achieve the following:
 - Estimate Multi-Gaussian Expansion (MGE) for optimal GALFIT utilization.
 - Generate simulated galaxy data images to refine your modeling techniques.
 
+
+
+Version 1.0.0
+===============
+
+
+GALFITools: A library for GALFIT
+
+GALFITools constitutes a collection of specialized routines to facilitate
+the utilization of GALFIT. The toolkit encompasses a spectrum of routines,
+spanning from mask creation, Point Spread Function (PSF) creation,
+Multi-gaussian expansion, sky computation, initial parameter estimation,
+to the comprehensive analysis of GALFIT's output
+
+
+
+Version 1.7.0
+===============
+
+GALFITools: A library for GALFIT
+
+GALFITools constitutes a collection of specialized routines to facilitate
+the utilization of GALFIT. The toolkit encompasses a spectrum of routines,
+spanning from mask creation, Point Spread Function (PSF) creation,
+Multi-gaussian expansion, sky computation, initial parameter estimation,
+to the analysis of GALFIT's output
+
+In this release:
+
+    fitlog2csv function was added to convert the fit.log file to a CSV file
+    Computation of sky with SkyDs9 and SkyRing includes an option to remove the top 80% and bottom 20% of pixel values
+    checkFile command added to check GALFIT file before run
+    mge2galfit reads directly from GALFIT file to reduce arg parsing options
+    creates curve of growth plot
+    minor changes in names and plots
+    bugs fixed
+
+
+Version 1.7.7
+===============
+
+
+GALFITools: A library for GALFIT
+
+GALFITools constitutes a collection of specialized routines to facilitate
+the utilization of GALFIT. The toolkit encompasses a spectrum of routines,
+spanning from mask creation, Point Spread Function (PSF) creation,
+Multi-gaussian expansion, sky computation, initial parameter estimation,
+to the analysis of GALFIT's output
+
+In this release:
+
+    getPeak function was added to indicate the coordinates of the galaxy's pixel with the highest count value
+    skyring function now writes a fits file to indicate the ring where the sky was computed
+    bugs fixed
+
+
+
+
+Version 1.15.0
+================
+
+GALFITools: A library for GALFIT
+
+GALFITools constitutes a collection of specialized routines to facilitate
+the utilization of GALFIT. The toolkit encompasses a spectrum of routines,
+spanning from mask creation, Point Spread Function (PSF) creation,
+Multi-gaussian expansion, sky computation, initial parameter estimation,
+to the analysis of GALFIT's output
+
+In this release:
+
+    The functions have been documented.
+    Console scripts return results in units of pixels and arc sec
+    New function GetSersic to estimate initial parameters for GALFIT input file
+    New function imarith to make arithmetic operations on images
+    New options for mge2galfit: position, ellipticity, and angular position
+    New Function makePSF to create PSF images using multi gaussian expansion
+    New function getKappa2 as an alternative to getKappa to estimate the curvature of a function
+    New function getBarSize to estimate the bar size
+    Bugs fixed
 
 
 
