+++ CONTRIBUTING.rst
@@ -125,7 +125,6 @@ Clone the repository
 
    to be able to import the package under development in the Python REPL.
 
-   .. todo:: if you are not using pre-commit, please remove the following item:
 
 #. Install |pre-commit|_::
 
@@ -277,7 +276,6 @@ on PyPI_, the following steps can be used to release a new version for
 
 
 .. <-- strart -->
-.. todo:: Please review and change the following definitions:
 
 .. |the repository service| replace:: GitHub
 .. |contribute button| replace:: "Create pull request"
