+++ .github/workflows/run-test.yml
@@ -19,7 +19,7 @@ jobs:
       fail-fast: false
       matrix:
         os: ["ubuntu-latest", "macos-latest", "windows-latest"]
-        python-version: ["3.9", "3.10", "3.11"]
+        python-version: ["3.9", "3.10", "3.11", "3.12"]
 
     steps:
       - uses: actions/checkout@v4
@@ -31,15 +31,13 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           python -m pip install tox
-          #python -m pip install nox
       - name: List installed packages
         run: pip list
       - name: Run tests with Tox 
-        run: |
-          tox -e default
-          #    nox -s tests-${{ matrix.python-version }}
-      # You only need to upload code coverage once to codecov unless you have a
-      # more complex build that you need coverage for.
-      #  - name: Upload coverage to Codecov
-      #  if: ${{ matrix.os == 'ubuntu-latest' &&  matrix.python-version == '3.10'}}
-      #  uses: codecov/codecov-action@v3
+        run: tox -e default
+      # Uncomment and configure the following step if you want to upload coverage to Codecov
+      # - name: Upload coverage to Codecov
+      #   if: ${{ matrix.os == 'ubuntu-latest' && matrix.python-version == '3.10' }}
+      #   uses: codecov/codecov-action@v3
+
+
