+++ README.rst
@@ -98,7 +98,7 @@ the commands in the console.
 
 In the example below, the command *getReComp* is executed in a terminal 
 using the GALFIT output file 
-`galfit.01 <img/galfit.01>`__  to calculate the effective radius of a model 
+`galfit.01 <docs/galfit.01>`__  to calculate the effective radius of a model 
 with three Sersic components:
 
 ::
