+++ .github/workflows/run-test.yml
@@ -4,7 +4,7 @@ on:
   pull_request:
   push:
     branches:
-      - main
+      - master
 
 # Use bash by default in all jobs
 defaults:
