+++ .github/workflows/run-test.yml
@@ -34,7 +34,7 @@ jobs:
       - name: List installed packages
         run: pip list
       - name: Run tests with Tox 
-        run: tox -e default -- -k "not  test_galfit"
+        run: tox -e default -- -k "not  test_galfit and not test_makePSF"
 
       # Uncomment and configure the following step if you want to upload coverage to Codecov
       # - name: Upload coverage to Codecov
