### 2025-02-27 #220 @canorve
Dear @SimonMolinsky, I have completed the documentation, which is hosted at [GALFITools Documentation](https://galfitools.readthedocs.io/en/latest/).

Additionally, GitHub Actions for automated tests have been implemented and tested.

I am now ready for feedback on both the documentation and the code
