### 2025-03-11 #220 @SimonMolinsky
Hi @canorve !

Sorry for the delay. I'm not as available as I would like to be! I've checked your docs and found one broken link under `here`—there is a typo in a link path, with an additional `/` after `.html`. Just remove it, and it will work fine.

```
First of all, install [GALFIT](https://users.obs.carnegiescience.edu/peng/work/galfit/galfit.html) if you haven’t done so. Check instructions [here](https://users.obs.carnegiescience.edu/peng/work/galfit/galfit.html/). Make sure that GALFIT can run in any path in your linux/macOS terminal.
```

GitHub Actions work, your documentation, especially the API description, is ready, and you have fulfilled all EiC checks. We can move forward with the review!
