### 2025-03-14 #220 @lwasser
hi colleagues! I noticed this has the astropy check on it. I"ve alerted the current astropy editors of that fact so it can be considered for astropy affiliation in the review.  Someone from the astropy editorial team will be in touch about the next steps. 🚀

@SimonMolinsky thank you so much for continuing to work on this submission 💓 You are so appreciated here.
