### 2025-04-14 #220 @coatless
@SimonMolinsky I'll handle this submission going forward. Thank you again for everything!

@canorve sorry for the delay in moving this forward. I've gone ahead and assigned @hamogu as the editor for the submission as it meets all of our requirements.

I will note that there are scientific similarities with PyAutoGalaxy that is also now under review; but, fundamentally the software is different in the underlying implementation and front facing uses.
