### 2025-04-15 #220 @hamogu
## Editor response to review:

Hi @SimonMolinsky ! Thanks for submitting your package to PyOpenSci. I'll be your editor - please feel free to reach out to me with any questions you might have. I'll now start to look for reviewers; usually that's the most time-consuming part of the process but I'll keep you updated as that goes along!

---

## Editor comments
:wave: Hi @Jammy2211 and @tepickering! Thank you for volunteering to review
for pyOpenSci!

- [ ] reviewer 3 (if applicable)

Please get in touch with any questions or concerns! Your review is due: <Insert deadline DATE HERE>
---

Reviewers: @Jammy2211 @tepickering
Due date: 11-Jun-2025
