### 2025-04-19 #220 @hamogu
@canorve : You indicated that you intend to have this package also submitted to JOSS by setting the tick mark in the submission. As you may know, the review we do will automatically be accepted as the review by JOSS; they will only do some editorial stuff, but not look at the package again. That means that the reviewers here also need to look at the `paper.md` that JOSS will publish. Could you point us to where we find that paper in your repro? (See links to JOSS documentation in the first post in this thread for details).
