### 2025-04-20 #220 @canorve
Thank you for pointing this out. I will review the JOSS documentation and upload the `paper.md` file to the repository at my earliest convenience. Once uploaded, I will notify you here.
