### 2025-05-02 #220 @canorve
Dear @hamogu,
Thank you for the update. I’ve received the message and appreciate the clarification. I’ll upload the paper.md as soon as possible.
