+++ paper/paper.bib
@@ -1,0 +1,98 @@
+
+
+---
+
+### 📚 `paper.bib`
+
+```bibtex
+@article{peng02,
+  author       = {Peng, C. Y. and Ho, L. C. and Impey, C. D. and Rix, H.-W.},
+  title        = {Detailed Structural Decomposition of Galaxy Images},
+  journal      = {The Astronomical Journal},
+  volume       = {124},
+  number       = {1},
+  pages        = {266--293},
+  year         = {2002},
+  doi          = {10.1086/340952}
+}
+
+@misc{anorve24,
+  author       = {Christopher Añorve},
+  title        = {{GALFITools v1.15.0}},
+  year         = {2024},
+  doi          = {10.5281/zenodo.13994492},
+  url          = {https://doi.org/10.5281/zenodo.13994492},
+  note         = {Zenodo}
+}
+
+
+@ARTICLE{peng10,
+       author = {{Peng}, Chien Y. and {Ho}, Luis C. and {Impey}, Chris D. and {Rix}, Hans-Walter},
+        title = "{Detailed Decomposition of Galaxy Images. II. Beyond Axisymmetric Models}",
+      journal = {\aj},
+     keywords = {galaxies: bulges, galaxies: fundamental parameters, galaxies: structure, techniques: image processing, techniques: photometric, Astrophysics - Cosmology and Nongalactic Astrophysics},
+         year = 2010,
+        month = jun,
+       volume = {139},
+       number = {6},
+        pages = {2097-2129},
+          doi = {10.1088/0004-6256/139/6/2097},
+archivePrefix = {arXiv},
+       eprint = {0912.0731},
+ primaryClass = {astro-ph.CO},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2010AJ....139.2097P},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@article{cappellari02,
+    author = {Cappellari, Michele},
+    title = {Efficient multi-Gaussian expansion of galaxies},
+    journal = {Monthly Notices of the Royal Astronomical Society},
+    volume = {333},
+    number = {2},
+    pages = {400-410},
+    year = {2002},
+    month = {06},
+    abstract = {We describe a simple, efficient, robust and fully automatic algorithm for the determination of a multi-Gaussian expansion (MGE) fit to galaxy images, to be used as a parametrization for the galaxy stellar surface brightness. In most cases the least-squares solution found by this method essentially corresponds to the minimax, constant relative error, MGE approximation of the galaxy surface brightness, with the chosen number of Gaussians. The algorithm is well suited to be used with multiple-resolution images (e.g. Hubble Space Telescope (HST) and ground-based images). It works orders of magnitude faster and is more accurate than currently available methods. An alternative, more computing-intensive, fully linear algorithm that is guaranteed to converge to the smallest χ2 solution is also discussed. Examples of MGE fits are presented for objects with HST or ground-based photometry, including galaxies with significant isophote twist.},
+    issn = {0035-8711},
+    doi = {10.1046/j.1365-8711.2002.05412.x},
+    url = {https://doi.org/10.1046/j.1365-8711.2002.05412.x},
+    eprint = {https://academic.oup.com/mnras/article-pdf/333/2/400/26001847/333-2-400.pdf},
+}
+
+
+@article{erwin15,
+doi = {10.1088/0004-637X/799/2/226},
+url = {https://dx.doi.org/10.1088/0004-637X/799/2/226},
+year = {2015},
+month = {jan},
+publisher = {The American Astronomical Society},
+volume = {799},
+number = {2},
+pages = {226},
[…]
