### 2025-05-15 #220 @canorve
Dear @hamogu, just to let you know that I have uploaded the paper.md file; it is located in the paper folder. I'm sorry for the delay.
