### 2025-05-21 #220 @hamogu
@tepickering and @Jammy2211 Thank you so much for volunteering to review this package. It should all be set for you to start looking. Review instructions are linked here https://github.com/pyOpenSci/software-submission/issues/220#issuecomment-2803592850 and please reach out to me with any questions! As is customary, I set the review due date for there weeks from now, but if you run into problems or need more time let me know and we'll work it out.
