### 2025-06-13 #220 @Jammy2211
I began my review, but after downloading the GALFIT debian64 kernel from the GALFIT home page, when I run it I get this error

`./galfit: error while loading shared libraries: libncurses.so.5: cannot open shared object file: No such file or directory`

I have tried other kernels, but no luck.

ChatGPT suggests I could down a route of changing my libncurses installation, but I have had issues with my windows subsystem for linux breaking when I start to change random libraries.

The alternative would be to try build GALFIT from source, but this would again require me to touch C libraries that I would rather not risk breaking.

I doubt there is a simple fix, so I see two ways forward:

1) I complete the review without actually running GALFITools, which is not ideal but is probably something I can get most of the way there.
2) You find another reviewer who (hopefully) has the hardware to run GALFIT.

If there is a fix which doesn't require me to risk running `sudo apt install something` I would be happy to try that.
