### 2025-06-13 #220 @hamogu
@canorve Do you have a suggestion? If not, I would also @Jammy2211 to review as good as you can (documentation and JOSS paper for example don't depend on actually installing it) and I will install it and do a partial review myself to be added to that.
