### 2025-06-13 #220 @tepickering
now that the AAS is over i'll finally have some time to devote to this. would it be helpful to set up a docker container with galfit and GALFITools installed? that would work around having to match the various dependencies on one's local machine, at least for review purposes.
