### 2025-06-13 #220 @canorve
I don't think there's an easier fix than running `sudo apt install libncurses-dev`
If it's helpful, GALFIT is only invoked when the _makePSF_ function is used; the rest of GALFITools does not depend on it.
