### 2025-06-17 #220 @Jammy2211
I attach my review below, as discussed above the vast majority of functionality did not actually require GALFIT and simply ran off of GALFIT results files, so I was able to go through the majority of code without issue.

- [x ] As the reviewer I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

- [x ] **Installation instructions:** for the development version of the package and any non-standard dependencies in README.

  - [x ] Tests cover essential functions of the package and a reasonable range of inputs and conditions.
- [x ] **Continuous Integration:** Has continuous integration setup (We suggest using Github actions but any CI platform is acceptable for review)

3

---

- My main critique is that the package assumes the user has a lot of Astronomy experience and knowledge in terms of how its documented and phrased.

For example, the documentation for the MGE reads as follows:

```
Routines that use the Multi-Gaussian Expansion.

mge2galfit fits multi-gaussian expansion of Cappellari (2002) and formats to GALFIT

positional arguments:
  GalfitFile            GALFIT file to obtain the header options
  Ds9regFile            the DS9 ellipse region file containing the galaxy

options:
  -h, --help            show this help message and exit
  -t, --twist           uses twist option for mge
  -c, --center          uses the center given in DS9 region file,otherwise it will found the x,y peak within DS9
                        ellipse
  -p PSF, --psf PSF     the value of PSF sigma
  -gas, --gauss         uses gauss function for galfit file
  -fser, --freeser      leaves the sersic index as a free parameter to fit
  -fsk, --freesky       leaves the sky as a free parameter to fit
  -ng NUMGAUSS, --numgauss NUMGAUSS
SbProf creates a surface brightness profile from a ellipse ds9 region

positional arguments:
  Image                 image fits file
  Ds9Region             Ds9 ellipse region file

options:
  -h, --help            show this help message and exit
  -q AXRAT, --axrat AXRAT
                        axis ratio
  -pa ANGLE, --angle ANGLE
                        angular position (same as GALFIT)
  -mz MGZPT, --mgzpt MGZPT
                        Magnitud zero point
  -m MASK, --mask MASK  mask fits file
  -s SKY, --sky SKY     sky value. Default = 0
  -p PLATE, --plate PLATE
                        plate scale
  -o OUTPUT, --output OUTPUT
                        output file
  -c, --center          uses the center given in DS9 region file,otherwise it will found the x,y
                        peak within DS9 ellipse
  -rx RANX RANX, --ranx RANX RANX
                        provide a range for x-axis: xmin - xmax
  -ry RANY RANY, --rany RANY RANY
                        provide a range for y-axis: ymin - ymax
  -lx, --logx           turn the X-axis to logarithm
  -px, --pix            turn the top x-axis in pixels
  -g, --grid            display a grid in the plot
  -r RAD, --rad RAD     value for a vertical line to add into the plot
  -r2 RAD2, --rad2 RAD2
                        value for a second vertical line to add into the plot
```

The documentation above, as far as I can tell, is essentially assuming one has used `mgefit` before and knows what a Multi Gaussian Expansion is.

For users who are not already familiar with a specific set of Astronomy tools (including things like DS9) I think it is quite difficult for one to get their head around what the correct inputs of GALFITools are in many cases.

Another example is this one:

```
getKappa2 gets the kappa radius from a set of Sersics using an alternative method to getKappa

positional arguments:
  GalfitFile            Galfit File containing the Sersics or gaussians components

options:
  -h, --help            show this help message and exit
  -d DIS, --dis DIS     Maximum distance among components
  -n NUMCOMP, --numcomp NUMCOMP
                        Number of component where it'll obtain center of all components, default = 1
  -a ANGLE, --angle ANGLE
                        Angle of the major axis of the galaxy. Default= it will take the angle of the
                        last components
  -p, --plot            makes plot of double derivative vs. radius
  -rx RANX RANX, --ranx RANX RANX
                        x-axis range to search for the Break radius: xmin - xmax
```

I don't know what the kappa radius is, I dont know why there is an alternative method to compute it and I don't know what that method does.

I am assuming this package is aimed at people who use a specific set of tools with known conventions and definitions within Astronomy, and that the package essentially acts as a streamlined interface for performing those calculations.

This may be fine and within the scope of a standalone software package, but for someone who only touched GALFIT a few times 10 years ago and hasn't used tools like DS9, I found it difficult to understand what a lot of the functionality does.

I am not sure what to suggest, on the one hand the documentation could be expanded to make this more accessible to a user who is unfamiliar with many of these tools. On the other hand, the package is targeting those with familiarity so the concise nature of the docs may be fine. Nevertheless, I raise here my experience and will see what others think.

- I can’t see the statement of need in the docs, only the JOSS paper has one, maybe move it to here: https://galfitools.readthedocs.io/en/latest/index.html

- Please include the supported Python versions (e..g. 3.10 - 3.12 ?) in the installation instructions.

- The link to “install GALFIT” on the installation page goes to the FAQ, which has no installation instructions. Link here instead: https://users.obs.carnegiescience.edu/peng/work/galfit/galfit.html

- Instructions on needing to install pytest-cov for unit testing.
