### 2025-06-17 #220 @tepickering
My review was done using an Apple M1 platform running Sequoia 15.5. The `galfit` binary for "Mavericks" works fine using Apple's Rosetta x86_64 emulation and I was able to run through the `galfit` examples.

3
---

- I found the statement of need to be sufficient because the package has a very specific target audience: the users of `galfit`. As such, some prior knowledge of `galfit` and its capabilities can be assumed. A significant background in astronomy and astronomical image analysis is a prerequisite for effectively using `galfit` and interpreting its results. The same prerequisite applies to this package.
- The link to the datafile used for the example, `galfit.01`, points to a file within the repository tree. As it's currently linked, it is a multi-step process to download into a usable format. One option would be to use a link to the raw file, e.g. [https://raw.githubusercontent.com/canorve/GALFITools/refs/heads/master/docs/galfit.01](https://raw.githubusercontent.com/canorve/GALFITools/refs/heads/master/docs/galfit.01), which would facilitate right-click->"save as", or include the file within the documentation tree and serve from RTD.
- The links to the "Module Reference", "Usage guide", and "API Reference" are kinda buried at the bottom of the documentation index page.  It would be nice to feature them more prominently so they're easier to find, e.g. in the "Documentation" section which currently only links recursively back to the main page.
- The README includes instructions for installation via `pip` and the instructions for installing a development version are in the documentation under "Contributing". I think this is fine and allows the developer instructions to be more detailed.
- Several requested badges are missing.
- I tested locally under python 3.13 and everything passed. I would recommend adding 3.13 to the CI test matrix.

My comments are all relatively minor and overall the package works very well for me. It's something I wish I had access to back when I was working on galaxy surface photometry long before even `galfit` existed...
