### 2025-08-13 9c24318
getMeRad function added
