+++ setup.cfg
@@ -105,6 +105,7 @@ console_scripts =
      getKappa = galfitools.shell.commands_galout:mainKappa
      getKappa2 = galfitools.shell.commands_galout:mainKappa2
      getReComp = galfitools.shell.commands_galout:mainGetReComp
+     getMeRad = galfitools.shell.commands_galout:mainGetMeRad
      getSlope = galfitools.shell.commands_galout:maingetSlope
      getN = galfitools.shell.commands_galout:maingetN
      getMissLight = galfitools.shell.commands_galout:mainMissingLight
