### 2025-08-14 644ffaf
getMeRad function computes mean surface brightness at any radius
