### 2025-08-14 92f0452
new function getBT to compute bulge_total ratio
