+++ setup.cfg
@@ -108,6 +108,7 @@ console_scripts =
      getMeRad = galfitools.shell.commands_galout:mainGetMeRad
      getSlope = galfitools.shell.commands_galout:maingetSlope
      getN = galfitools.shell.commands_galout:maingetN
+     getBT = galfitools.shell.commands_galout:maingetBT
      getMissLight = galfitools.shell.commands_galout:mainMissingLight
      getBulgeRad = galfitools.shell.commands_galout:mainGetBulgeRad
      showCube = galfitools.shell.commands_galout:mainShowCube
