### 2025-08-14 ae96895
removed bug for getBT function
