### 2025-08-14 fc6afae
getBT functions updated
