+++ src/galfitools/galout/getBT.py
@@ -140,152 +140,6 @@ def bulge_to_total(mag_bulge, mag_disk, mag_bar=99):
     return bulge_total
 
 
-class GetN:
-    """Computes the Sersic index from photometric parameters
-
-
-
-    Methods
-    -------
-    GalNs : computes the Sersic index from effective radius and
-            other fraction of light radius.
-
-    MeMeanMe : computes the Sersic index from surface brightness
-                at Re and mean surface brightness at Re.
-
-
-    """
-
-    def GalNs(self, EffRad: float, EffRadfrac: list, F: list):
-
-        sers = np.array([])
-
-        for idx, f in enumerate(F):
-
-            n = self.ReRfrac(EffRad, EffRadfrac[idx], f)
-
-            sers = np.append(sers, n)
-
-        return sers
-
-    def MeMeanMe(self, me: float, meanme: float) -> float:
-
-        a = 0.1
-        b = 12
-
-        Sersic = self.solveSer(a, b, me, meanme)
-
-        return Sersic
-
-    def solveSer(self, a: float, b: float, me: float, meanme: float) -> float:
-        "return the sersic index. It uses Bisection"
-
-        try:
-            N = bisect(self.funMeMeanMe, a, b, args=(me, meanme))
-        except Exception:
-            print("unable to solve equation in the given range. Setting n to 99")
-            N = 99
-
-        return N
-
-    def funMeMeanMe(self, n: float, me: float, meanme: float) -> float:
-
-        # k = gammaincinv(2 * n, 0.5)
-
-        fn = self.Fn(n)
-
-        result = me - meanme - 2.5 * np.log10(fn)
-
-        return result
-
-    def Fn(self, n: float) -> float:
-
-        k = gammaincinv(2 * n, 0.5)
-
-        fn = ((n * np.exp(k)) / (k ** (2 * n))) * gamma(2 * n)
-
-        return fn
-
-    def ReRfrac(self, Re: float, Rfrac: float, frac: float) -> float:
-
-        a = 0.1
-        b = 12
-
-        Sersic = self.solveSerRe(a, b, Re, Rfrac, frac)
-
-        return Sersic
-
-    def solveSerRe(
-        self, a: float, b: float, Re: float, Rfrac: float, frac: float
-    ) -> float:
-        "return the sersic index. It uses Bisection"
-
-        try:
-            N = bisect(self.funReRfrac, a, b, args=(Re, Rfrac, frac))
-        except Exception:
-            print("unable to solve equation in the given range. Setting n to 99")
-            N = 99
-
-        return N
-
-    def funReRfrac(self, n: float, Re: float, Rfrac: float, frac: float) -> float:
-
-        k = gammaincinv(2 * n, 0.5)
-
-        x = k * (Rfrac / Re) ** (1 / n)
-
-        result = frac * gamma(2 * n) - gamma(2 * n) * gammainc(2 * n, x)
-
-        return result
-
-    def MeM0(self, me: float, m0: float) -> float:
-        """Uses me and m0 to compute n. It is not very realiable
-        and takes longer than the other two methods"""
-        a = 0
-        b = 45
-
-        K = self.solveKm0(a, b, me, m0)
-
-        a = 0.2
-        b = 40
-
-        Sersic = self.solveSerK(a, b, K)
-
-        return Sersic
-
-    def solveKm0(self, a: float, b: float, me: float, m0: float) -> float:
-        "return the sersic index. It uses Bisection"
-
-        try:
-            K = bisect(self.funMeM0, a, b, args=(me, m0))
-        except Exception:
-            print("solution not found for the given range")
-            K = 0
-
-        return K
-
-    def funMeM0(self, K: float, me: float, m0: float) -> float:
-
-        result = me - m0 - 2.5 * K / np.log(10)
-
-        return result
-
[…]
