+++ README.rst
@@ -54,11 +54,9 @@ expansion (MGE) fitting, as well as computation of sky background
 and other pertinent photometric variables.
 
 
-
 **NOTE:** GALFITools assumes the user has a basic knowledge of GALFIT and SaoImage DS9.  
-How to use GALFIT is found in the official `GALFIT <GALFIT2_>`_ webpage.  
-The DS9 webpage can be found here_ and a basic tutorial `here <here2_>`_.
-
+How to use GALFIT is found in the official `GALFIT <GALFIT2>`_ webpage.  
+The DS9 webpage can be found `here <here>`_ and a basic tutorial `here <here2>`_.
 
 .. _GALFIT: https://users.obs.carnegiescience.edu/peng/work/galfit/GFAQ.html
 .. _GALFIT2: https://users.obs.carnegiescience.edu/peng/work/galfit/galfit.html
