+++ README.rst
@@ -43,6 +43,9 @@
    :target: https://github.com/canorve/galfitools/actions/workflows/run-test.yml
    :alt: Build Status
 
+.. image:: https://github.com/canorve/galfitools/actions/workflows/run-test.yml/badge.svg
+   :target: https://github.com/canorve/galfitools/actions/workflows/run-test.yml
+   :alt: Build Status
 
 ============
 GALFITools
