### 2025-08-18 b0bc8bb
readme updated and coverage implemented
