+++ tox.ini
@@ -16,9 +16,10 @@ passenv =
     HOME
 extras =
     testing
+deps = 
+    pytest-cov
 commands =
-    pytest {posargs}
-
+    pytest --cov=. --cov-report=xml {posargs}
 
 # # To run `tox -e lint` you need to make sure you have a
 # # `.pre-commit-config.yaml` file. See https://pre-commit.com
