+++ .github/workflows/run-test.yml
@@ -37,8 +37,11 @@ jobs:
         run: tox -e default -- -k "not  test_galfit and not test_makePSF"
 
       # Uncomment and configure the following step if you want to upload coverage to Codecov
-      # - name: Upload coverage to Codecov
-      #   if: ${{ matrix.os == 'ubuntu-latest' && matrix.python-version == '3.10' }}
-      #   uses: codecov/codecov-action@v3
+      - name: Upload coverage to Codecov
+        if: ${{ matrix.os == 'ubuntu-latest' && matrix.python-version == '3.10' }}
+        uses: codecov/codecov-action@v4
+        with:
+          file: ./coverage.xml
+          fail_ci_if_error: true
 
 
