+++ .github/workflows/run-test.yml
@@ -3,12 +3,12 @@ name: Pytest unit/integration
 on:
   pull_request:
   push:
-    branches:
-      - master
+    branches: [ master ]
 
 permissions:
   contents: read
   pull-requests: write
+  id-token: write
 
 # Use bash by default in all jobs
 defaults:
@@ -27,6 +27,8 @@ jobs:
 
     steps:
       - uses: actions/checkout@v4
+        with:
+            fetch-depth: 0
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v4
         with:
@@ -40,13 +42,20 @@ jobs:
       - name: Run tests with Tox 
         run: tox -e default -- -k "not  test_galfit and not test_makePSF"
 
+      - name: Inspect coverage.xml
+        if: ${{ matrix.os == 'ubuntu-latest' && matrix.python-version == '3.10' }}
+        run: |
+          ls -l coverage.xml || (echo "coverage.xml not found" && exit 1)
+          head -n 5 coverage.xml
+
+
       # Uncomment and configure the following step if you want to upload coverage to Codecov
       - name: Upload coverage to Codecov
         if: ${{ matrix.os == 'ubuntu-latest' && matrix.python-version == '3.10' }}
         uses: codecov/codecov-action@v4
         with:
+          use_oidc: true
           file: ./coverage.xml
           fail_ci_if_error: true
-          fetch-depth: 0
 
 
