+++ .github/workflows/run-test.yml
@@ -67,7 +67,7 @@ jobs:
 
       - name: Upload coverage to Codecov
         if: ${{ matrix.os == 'ubuntu-latest' && matrix.python-version == '3.10' }}
-        uses: codecov/codecov-action@v4
+        uses: codecov/codecov-action@v3
         env:
           # Make the CLI see the correct repo
           CODECOV_SLUG: canorve/GALFITools
