+++ .github/workflows/run-test.yml
@@ -59,15 +59,18 @@ jobs:
       #    fail_ci_if_error: true
 
 
-      - name: Show repo context seen by action
-        if: ${{ matrix.os == 'ubuntu-latest' && matrix.python-version == '3.10' }}
-        run: |
-          echo "GITHUB_REPOSITORY=$GITHUB_REPOSITORY"
-          echo "CODECOV_SLUG=$CODECOV_SLUG"
+          #- name: Show repo context seen by action
+          #if: ${{ matrix.os == 'ubuntu-latest' && matrix.python-version == '3.10' }}
+          #run: |
+          #echo "GITHUB_REPOSITORY=$GITHUB_REPOSITORY"
+          #echo "CODECOV_SLUG=$CODECOV_SLUG"
+
+
+          #if: ${{ matrix.os == 'ubuntu-latest' && matrix.python-version == '3.10' }}
 
       - name: Upload coverage to Codecov
-        if: ${{ matrix.os == 'ubuntu-latest' && matrix.python-version == '3.10' }}
-        uses: codecov/codecov-action@v3
+        if: ${{ success() && matrix.os == 'ubuntu-latest' && matrix.python-version == '3.10' }}
+        uses: codecov/codecov-action@v4
         env:
           # Make the CLI see the correct repo
           CODECOV_SLUG: canorve/GALFITools
