+++ .github/workflows/run-test.yml
@@ -24,6 +24,9 @@ jobs:
       matrix:
         os: ["ubuntu-latest", "macos-latest"]
         python-version: ["3.9", "3.10", "3.11", "3.12"]
+        exclude:
+          - os: ubuntu-latest
+            python-version:"3.10"
 
     steps:
       - uses: actions/checkout@v4
@@ -71,19 +74,12 @@ jobs:
       - name: Upload coverage to Codecov
         if: ${{ success() && matrix.os == 'ubuntu-latest' && matrix.python-version == '3.10' }}
         uses: codecov/codecov-action@v4
-        env:
-          # Make the CLI see the correct repo
-          CODECOV_SLUG: canorve/GALFITools
-          GITHUB_REPOSITORY: canorve/GALFITools
         with:
-          use_oidc: true
-          git_service: github
           slug: canorve/GALFITools
           files: ./coverage.xml
           override_branch: master
           override_commit: ${{ github.sha }}
           root_dir: .
-          use_legacy_upload_endpoint: true
           fail_ci_if_error: true
           verbose: true
 
