+++ .github/workflows/run-test.yml
@@ -26,7 +26,7 @@ jobs:
         python-version: ["3.9", "3.10", "3.11", "3.12"]
         exclude:
           - os: ubuntu-latest
-            python-version:"3.10"
+            python-version: "3.10"
 
     steps:
       - uses: actions/checkout@v4
