### 2025-08-19 a1509ab
readme and setup updated
