+++ setup.cfg
@@ -40,7 +40,11 @@ classifiers =
     Natural Language :: Spanish
     Natural Language :: English
     Topic :: Scientific/Engineering :: Astronomy
-
+    Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
+    Programming Language :: Python :: 3.11
+    Programming Language :: Python :: 3.12
+    Programming Language :: Python :: 3.13
 
 [options]
 zip_safe = False
