+++ .github/workflows/run-test.yml
@@ -82,6 +82,7 @@ jobs:
           root_dir: .
           fail_ci_if_error: true
           verbose: true
+          token: ${{ secrets.CODECOV_TOKEN }}
 
 
 
