+++ setup.cfg
@@ -55,7 +55,7 @@ package_dir =
 
 # Require a min/specific Python version (comma-separated conditions)
 # python_requires = >=3.8
- python_requires = >=3.9
+python_requires = >=3.9
 
 # Add here dependencies of your project (line-separated), e.g. requests>=2.2,<3.0.
 # Version specifiers like >=2.2,<3.0 avoid problems due to API changes in
