+++ .github/workflows/run-test.yml
@@ -71,18 +71,34 @@ jobs:
 
           #if: ${{ matrix.os == 'ubuntu-latest' && matrix.python-version == '3.10' }}
 
+          #- name: Upload coverage to Codecov
+          #if: ${{ success() && matrix.os == 'ubuntu-latest' && matrix.python-version == '3.10' }}
+          #uses: codecov/codecov-action@v4
+          #with:
+          #slug: canorve/GALFITools
+          #files: ./coverage.xml
+          #override_branch: master
+          #override_commit: ${{ github.sha }}
+          #root_dir: .
+          #fail_ci_if_error: true
+          #verbose: true
+          #token: ${{ secrets.CODECOV_TOKEN }}
+
+        
       - name: Upload coverage to Codecov
-        if: ${{ success() && matrix.os == 'ubuntu-latest' && matrix.python-version == '3.10' }}
+        if: ${{ success()
+              && github.event_name == 'push'
+              && github.ref == 'refs/heads/master'
+              && matrix.os == 'ubuntu-latest'
+              && matrix.python-version == '3.11' }}
         uses: codecov/codecov-action@v4
         with:
-          slug: canorve/GALFITools
           files: ./coverage.xml
+          token: ${{ secrets.CODECOV_TOKEN }}   # prevents throttling
+          slug: canorve/GALFITools
           override_branch: master
           override_commit: ${{ github.sha }}
           root_dir: .
-          fail_ci_if_error: true
+          fail_ci_if_error: false
           verbose: true
-          token: ${{ secrets.CODECOV_TOKEN }}
-
-
 
