+++ tox.ini
@@ -17,9 +17,24 @@ passenv =
 extras =
     testing
 deps = 
-    pytest-cov
+    pytest>=7
+    pytest-cov>=5
 commands =
-    pytest --cov=. --cov-report=xml {posargs}
+    pytest --cov=galfitools --cov-report=xml {posargs}
+
+#pytest -p pytest_cov --cov=galfitools --cov-report=xml {posargs}
+   
+[run]
+source = galfitools
+
+
+#    python -m coverage erase
+#    python -m coverage run --source=galfitools -m pytest {posargs}
+#    python -m coverage xml
+
+
+
+# pytest --cov=. --cov-report=xml {posargs}
 
 # # To run `tox -e lint` you need to make sure you have a
 # # `.pre-commit-config.yaml` file. See https://pre-commit.com
