+++ tox.ini
@@ -17,15 +17,18 @@ passenv =
 extras =
     testing
 deps = 
-    pytest>=7
-    pytest-cov>=5
+    pytest
+    pytest-cov
 commands =
-    pytest --cov=galfitools --cov-report=xml {posargs}
+    pytest --cov=src/galfitools --cov-report=xml {posargs}
 
-#pytest -p pytest_cov --cov=galfitools --cov-report=xml {posargs}
+
+#pytest --cov=galfitools --cov-report=xml --cov-config=setup.cfg {posargs}
+# pytest --cov=galfitools --cov-report=xml {posargs}
+# pytest -p pytest_cov --cov=galfitools --cov-report=xml {posargs}
    
-[run]
-source = galfitools
+#[run]
+#source = galfitools
 
 
 #    python -m coverage erase
