+++ docs/statement.rst
@@ -1,0 +1,59 @@
+
+.. _statement:
+
+**GALFITools: Statement of need**
+============================
+
+
+.. contents::
+   :depth: 3
+..
+
+.. image:: https://img.shields.io/pypi/v/GALFITools.svg
+    :alt: PyPI-Server
+    :target: https://pypi.org/project/GALFITools/
+
+.. image:: https://img.shields.io/badge/-PyScaffold-005CA0?logo=pyscaffold
+    :alt: Project generated with PyScaffold
+    :target: https://pyscaffold.org/
+
+.. image:: https://zenodo.org/badge/DOI/10.5281/zenodo.8248042.svg
+   :target: https://doi.org/10.5281/zenodo.8248042
+
+
+
+---------------------
+
+** Statement of need**
+---------------------
+
+
+
+
+The analysis of galaxy morphology through image fitting is a fundamental task 
+in extragalactic astronomy. GALFIT [@peng02; @peng10] is a well-established tool that performs parametric 
+two-dimensional modeling of galaxy surface brightness profiles. However, GALFIT does 
+not provide utilities for related tasks such as generating mask images, selecting 
+stars for PSF modeling, estimating initial fit parameters, or interpreting its output 
+formats in a structured way. These tasks are typically handled manually or with ad hoc
+scripts, which can reduce reproducibility and increase the barrier for new users.
+
+**GALFITools** [@anorve24] addresses this gap by offering a cohesive suite of Python-based tools 
+that extend the functionality of GALFIT. It facilitates the full modeling pipeline, 
+from input preparation to result interpretation. This includes routines to construct 
+PSFs, galaxy modeling via Multi-Gaussian Expansion [MGE, @cappellari02], estimate sky backgrounds,  simulate 
+galaxies, estimate initial parameters, generate diagnostic plots, model visualization, 
+and compute photometric parameters from multiple Sérsic
+components such as effective radius, bar size, Sérsic index,  (among others). 
+These functionalities are available as command-line tools, in line with GALFIT's command-line 
+interface, but can also be imported as Python modules.
+
+While other packages such as `IMFIT` [@erwin15] and `ProFit` [@robotham16] provide similar modeling capabilities, 
+GALFIT remains one of the most widely used tools in the astronomical community for galaxy 
+image modeling, supported by an extensive user base and numerous legacy workflows. GALFITools 
+is not a replacement for GALFIT, but a complementary toolset specifically tailored to its 
+ecosystem. By reducing the technical overhead and providing automation, GALFITools supports 
+researchers in conducting large-scale studies of galaxy structure and photometry with 
+improved efficiency and reproducibility.
+
+
