+++ docs/statement.rst
@@ -3,10 +3,10 @@
 
 
 
----------------------
+============
+Statement of need
+============
 
-**Statement of need**
----------------------
 
 
 
