+++ docs/index.rst
@@ -42,7 +42,6 @@ Indices and tables
 ==================
 
 * :ref:`genindex`
-* :ref:`readme`
 * :ref:`usage`
 * :ref:`api`
 * :ref:`modindex`
