+++ docs/concepts.rst
@@ -8,32 +8,36 @@ the package more accessible for users who may be less familiar with
 astronomical jargon.
 
 
-.. _concept-surface-brightness:
 
-Surface brightness (SB)
------------------------
-Surface brightness is the flux received per unit area on the sky,
-commonly reported in magnitudes per square arcsecond (mag arcsec⁻²).
-It describes how light is distributed across a galaxy image.
+Software Tools
+--------------
 
-.. _concept-psf:
+.. _concept-galfit:
 
-Point–spread function (PSF)
----------------------------
-The point–spread function represents the response of the telescope and
-detector to a point source of light (such as a star). It describes the
-blurring of the image due to the instrument and atmosphere (for
-ground–based observations). GALFIT requires a PSF to deconvolve models.
+GALFIT
+------
+GALFIT is a widely used program for two–dimensional fitting of galaxy
+images (Peng et al. 2002). It models galaxies with analytic functions
+(e.g., Sérsic, exponential disks, Nuker profiles) convolved with the
+point–spread function (PSF). GALFIT is the external engine that
+GALFITools interfaces with.
 
-.. _concept-sersic:
+.. _concept-ds9:
 
-Sérsic profile and index
-------------------------
-A Sérsic profile is a mathematical function that describes how the
-brightness of a galaxy varies with radius. The **Sérsic index** (*n*)
-controls the shape of the profile: low *n* values represent disk–like,
-exponential profiles, while high *n* values represent more concentrated,
-bulge–like profiles.
+SaoImage DS9
+------------
+SaoImage DS9 is an astronomical imaging and visualization application.
+It is commonly used to inspect FITS images, define regions of interest,
+and interactively examine astronomical data.
+
+.. _concept-ds9-regions:
+
+DS9 regions
+-----------
+In DS9, a *region* is a user–defined geometric shape (circle, box,
+polygon, etc.) drawn on an image. Regions can mark sources, masks,
+or fitting boundaries, and can be saved to files that GALFIT or
+GALFITools can read.
 
 .. _concept-mask:
 
@@ -53,6 +57,74 @@ as galaxy magnitude, effective radius, axis ratio, and Sérsic index.
 Providing reasonable initial values helps GALFIT converge to the correct
 solution.
 
+
+
+
+
+
+Image Properties
+----------------
+
+.. _concept-star-image:
+
+Star (in an image)
+------------------
+In the context of astronomical images, a *star* usually appears as a
+point–like source broadened by the PSF. Stars are often used to
+construct PSFs or to calibrate the image.
+
+.. _concept-psf:
+
+Point–spread function (PSF)
+---------------------------
+The point–spread function represents the response of the telescope and
+detector to a point source of light (such as a star). It describes the
+blurring of the image due to the instrument and atmosphere (for
+ground–based observations). GALFIT requires a PSF to deconvolve models.
+
+.. _concept-fwhm:
+
+Full Width at Half Maximum (FWHM)
+---------------------------------
+FWHM is a measure of the width of a profile. It is the distance between
+the two points on the profile where the value falls to half of its
+maximum. In astronomy, the FWHM of a stellar image provides an estimate
+of the PSF size and image resolution.
+
+.. _concept-surface-brightness:
+
+Surface brightness (SB)
+-----------------------
+Surface brightness is the flux received per unit area on the sky,
+commonly reported in magnitudes per square arcsecond (mag arcsec⁻²).
+It describes how light is distributed across a galaxy image.
+
+.. _concept-photometry:
+
+Photometry
+----------
+Photometry is the measurement of fluxes or magnitudes of astronomical
+objects. It can be performed with apertures, PSF fitting, or model
+fitting methods such as GALFIT.
+
[…]
