+++ docs/concepts.rst
@@ -9,8 +9,9 @@ astronomical jargon.
 
 
 
+--------------------
 **Software Tools**
---------------
+--------------------
 
 .. _concept-galfit:
 
@@ -139,8 +140,9 @@ surface brightness of a galaxy changes with radius. Profiles are often
 used to characterize galaxy structure and to fit analytic models.
 
 
+--------------------
 **Image Properties**
-----------------
+--------------------
 
 .. _concept-star-image:
 
@@ -262,53 +264,56 @@ objects, while smaller values indicate more elongated shapes.
 
 
 
+--------------------
+
 **Galaxy Components**
------------------
+---------------------
 
 .. _concept-bulge:
 
-Bulge
------
+**Bulge**
+
 The bulge is the central, spheroidal component of a galaxy. It is
 generally more concentrated and has higher surface brightness than the
 surrounding disk.
 
 .. _concept-disk:
 
-Disk
-----
+**Disk**
 The disk is the flattened, rotating component of a galaxy, typically
 hosting spiral arms and ongoing star formation. Its brightness profile
 is often well described by an exponential law.
 
 .. _concept-bar:
 
-Galactic bar
-------------
+**Galactic bar**
+
 A bar is an elongated structure of stars crossing the central region of
 a disk galaxy. Bars redistribute angular momentum and can drive gas
 inflows toward the galaxy center.
 
 .. _concept-bt:
 
-Bulge–to–total luminosity ratio (B/T)
--------------------------------------
+
+**Bulge–to–total luminosity ratio (B/T)**
+
 The bulge–to–total luminosity ratio is the fraction of a galaxy’s total
 light that comes from the bulge compared to the sum of bulge and disk.
 It is commonly used to quantify galaxy morphology.
 
 .. _concept-effective-radius:
 
-Effective radius (Re)
----------------------
+**Effective radius (Re)**
+
 The effective radius is the radius of a circular aperture that contains
 half of the total light of a galaxy or model component. It is a standard
 measure of galaxy size.
 
 
 
+------------------------------
 **Modeling and Mathematics**
-------------------------
+------------------------------
 
 .. _concept-sersic:
 
@@ -384,5 +389,6 @@ the mathematical definition.
 **Related GALFITools API**
 
 - :py:func:`galfitools.galout.getRads.getKappa`
+- :py:func:`galfitools.galout.getRads.getKappa2`
 
 
