+++ docs/concepts.rst
@@ -263,14 +263,14 @@ shape of a galaxy component. Values near 1 correspond to nearly circular
 objects, while smaller values indicate more elongated shapes.
 
 
-
 --------------------
 **Galaxy Components**
 ---------------------
 
 .. _concept-bulge:
 
-**Bulge**
+Bulge
+-------
 
 The bulge is the central, spheroidal component of a galaxy. It is
 generally more concentrated and has higher surface brightness than the
@@ -278,14 +278,17 @@ surrounding disk.
 
 .. _concept-disk:
 
-**Disk**
+Disk
+-------
+
 The disk is the flattened, rotating component of a galaxy, typically
 hosting spiral arms and ongoing star formation. Its brightness profile
 is often well described by an exponential law.
 
 .. _concept-bar:
 
-**Galactic bar**
+Galactic bar
+-----------
 
 A bar is an elongated structure of stars crossing the central region of
 a disk galaxy. Bars redistribute angular momentum and can drive gas
@@ -294,7 +297,8 @@ inflows toward the galaxy center.
 .. _concept-bt:
 
 
-**Bulge–to–total luminosity ratio (B/T)**
+Bulge–to–total luminosity ratio (B/T)
+-----------
 
 The bulge–to–total luminosity ratio is the fraction of a galaxy’s total
 light that comes from the bulge compared to the sum of bulge and disk.
@@ -302,7 +306,8 @@ It is commonly used to quantify galaxy morphology.
 
 .. _concept-effective-radius:
 
-**Effective radius (Re)**
+Effective radius (Re)
+-----------
 
 The effective radius is the radius of a circular aperture that contains
 half of the total light of a galaxy or model component. It is a standard
