+++ docs/concepts.rst
@@ -397,7 +397,11 @@ the mathematical definition.
 - :py:func:`galfitools.galout.getRads.getKappa`
 - :py:func:`galfitools.galout.getRads.getKappa2`
 
-**Related GALFITools CLI command**
+.. admonition:: Related GALFITools CLI commands
+   :class: seealso
 
-- :ref:`getKappa <routine-getKappa>`
-- :ref:`getKappa2 <routine-getKappa2>`
+   - :ref:`getKappa <routine-getKappa>`
+   - :ref:`getKappa2 <routine-getKappa2>`
+
+
+
