+++ docs/usage.rst
@@ -128,10 +128,12 @@ with GALFITools v1.11.0:
 **GALFIT Input**
 --------------------
 
+
 Routines that aid the GALFIT's user to
 prepare the necessary files for GALFIT input 
 
 
+.. _routine-getStar:
 **getStar** gets a image slice centered on the object peak
 
 ::
@@ -152,6 +154,7 @@ prepare the necessary files for GALFIT input
     -so SIGOUT, --sigout SIGOUT
                           the sigma image output
 
+.. _routine-initGal:
 **initGal** Creates GALFIT's input files with different initial parameters
 
 
@@ -185,6 +188,7 @@ prepare the necessary files for GALFIT input
        
 
 
+.. _routine-gtmakeMask:
 **gtmakeMask**  creates mask file from a SExtractor's catalog 
 
 ::
@@ -228,6 +232,7 @@ prepare the necessary files for GALFIT input
 
 
 
+.. _routine-maskDs9:
 **maskDs9**  creates (or modify) a mask image for GALFIT using DS9 regions 
 such as Boxes, Ellipses and Polygons
 
@@ -250,6 +255,7 @@ such as Boxes, Ellipses and Polygons
                           value of the border if it is different from zero
 
 
+.. _routine-maskSky:
 **maskSky** creates a mask image for GALFIT using original image and sky mean and sigma
 
 ::
@@ -275,6 +281,7 @@ such as Boxes, Ellipses and Polygons
     -bv BORVALUE, --borValue BORVALUE
                           value of the border if it is different from zero
 
+.. _routine-xy2fits:
 **xy2fits** code to convert ASCII x,y positions to FTIS mask
 
 ::
@@ -292,6 +299,7 @@ such as Boxes, Ellipses and Polygons
 
 
 
+.. _routine-checkFile:
 **checkFile** check that the parameters and file names inside the GALFIT input file are correct 
 
 ::
@@ -306,6 +314,7 @@ such as Boxes, Ellipses and Polygons
     -d DIS, --dis DIS  Maximum distance in pixels among components. Default = 10
 
 
+.. _routine-boxSize:
 **boxSize** computes the box size from a ds9 box region for galfit header option H) 
 
 ::
@@ -323,6 +332,7 @@ such as Boxes, Ellipses and Polygons
 
 
 
+.. _routine-getPeak:
 **getPeak**  Obtains the center, axis ratio and angular position from DS9 region
 
 ::
@@ -337,6 +347,7 @@ such as Boxes, Ellipses and Polygons
       -m MASK, --mask MASK  the mask file
 
 
+.. _routine-imarith:
 **imarith** makes arithmetic operations on image 
 
 ::
@@ -366,6 +377,7 @@ such as Boxes, Ellipses and Polygons
 
  
 
+.. _routine-getSersic:
 **getSersic** Its estimates and prints initial parameters for Sersic components. It
               addtion if proved options for single Sersic, bulge/disk and bulge/bar/disk
 
@@ -398,6 +410,7 @@ such as Boxes, Ellipses and Polygons
                           activated
 
 
+.. _routine-makePSF:
 **MakePSF** Makes a PSF model of a star using Multi Gaussian Expansion
 
 ::
@@ -427,10 +440,15 @@ such as Boxes, Ellipses and Polygons
 **GALFIT Output**
 -------------------
 
+
+
+
+
 Routines that computes photometric variables from 
 the surface brightness models fitted by GALFIT 
 
 
+.. _routine-getBreak:
 **getBreak** gets the break radius from a set of Sersics
 
 ::
@@ -460,6 +478,7 @@ the surface brightness models fitted by GALFIT
[…]
