+++ docs/concepts.rst
@@ -23,6 +23,19 @@ images (Peng et al. 2002). It models galaxies with analytic functions
 point–spread function (PSF). GALFIT is the external engine that
 GALFITools interfaces with.
 
+**Related GALFITools API**
+
+- :py:func:`galfitools.galout.getRads.getKappa`
+- :py:func:`galfitools.galout.getRads.getKappa2`
+
+
+.. admonition:: Related GALFITools CLI commands
+   :class: seealso
+
+   - :ref:`getKappa <routine-getKappa>`
+   - :ref:`getKappa2 <routine-getKappa2>`
+
+
 
 .. _concept-galfit-file:
 
@@ -37,6 +50,17 @@ radius, Sérsic index). After a run, the file can also record the final
 fitted parameters. GALFITools provides utilities to create and manage
 these files programmatically.
 
+**Related GALFITools API**
+
+- :py:func:`galfitools.galout.getRads.getKappa`
+- :py:func:`galfitools.galout.getRads.getKappa2`
+
+
+.. admonition:: Related GALFITools CLI commands
+   :class: seealso
+
+   - :ref:`getKappa <routine-getKappa>`
+   - :ref:`getKappa2 <routine-getKappa2>`
 
 
 Example of a GALFIT file
@@ -76,6 +100,17 @@ parameter is free or fixed during fitting.
   10) 45.0     1             # Position angle (degrees)
    Z) 0                      # Skip this model in output image? (no=0)
 
+**Related GALFITools API**
+
+- :py:func:`galfitools.galout.getRads.getKappa`
+- :py:func:`galfitools.galout.getRads.getKappa2`
+
+
+.. admonition:: Related GALFITools CLI commands
+   :class: seealso
+
+   - :ref:`getKappa <routine-getKappa>`
+   - :ref:`getKappa2 <routine-getKappa2>`
 
 
 .. _concept-ds9:
@@ -86,6 +121,18 @@ SaoImage DS9 is an astronomical imaging and visualization application.
 It is commonly used to inspect FITS images, define regions of interest,
 and interactively examine astronomical data.
 
+**Related GALFITools API**
+
+- :py:func:`galfitools.galout.getRads.getKappa`
+- :py:func:`galfitools.galout.getRads.getKappa2`
+
+
+.. admonition:: Related GALFITools CLI commands
+   :class: seealso
+
+   - :ref:`getKappa <routine-getKappa>`
+   - :ref:`getKappa2 <routine-getKappa2>`
+
 .. _concept-ds9-regions:
 
 DS9 regions
@@ -94,6 +141,19 @@ In DS9, a *region* is a user–defined geometric shape (circle, box,
 polygon, etc.) drawn on an image. Regions can mark sources, masks,
 or fitting boundaries, and can be saved to files that GALFITools 
 can read.
+
+**Related GALFITools API**
+
+- :py:func:`galfitools.galout.getRads.getKappa`
+- :py:func:`galfitools.galout.getRads.getKappa2`
+
+
+.. admonition:: Related GALFITools CLI commands
+   :class: seealso
+
+   - :ref:`getKappa <routine-getKappa>`
+   - :ref:`getKappa2 <routine-getKappa2>`
+
 
 .. _concept-mask:
 
@@ -104,6 +164,19 @@ fitting (for example, bright foreground stars, cosmic rays, or image
 defects). GALFITools includes utilities to create masks automatically or
 manually.
 
+**Related GALFITools API**
+
+- :py:func:`galfitools.galout.getRads.getKappa`
+- :py:func:`galfitools.galout.getRads.getKappa2`
+
+
+.. admonition:: Related GALFITools CLI commands
+   :class: seealso
+
+   - :ref:`getKappa <routine-getKappa>`
[…]
