+++ docs/concepts.rst
@@ -25,15 +25,15 @@ GALFITools interfaces with.
 
 **Related GALFITools API**
 
-- :py:func:`galfitools.galout.getRads.getKappa`
-- :py:func:`galfitools.galout.getRads.getKappa2`
+- :py:func:`galfitools.galin.MakePSF.makePSF`
+- :py:func:`galfitools.mge.mge2galfit.mge2gal`
 
 
 .. admonition:: Related GALFITools CLI commands
    :class: seealso
 
-   - :ref:`getKappa <routine-getKappa>`
-   - :ref:`getKappa2 <routine-getKappa2>`
+   - :ref:`makePSF <routine-makePSF>`
+   - :ref:`mge2galfit <routine-mge2galfit>`
 
 
 
@@ -52,15 +52,15 @@ these files programmatically.
 
 **Related GALFITools API**
 
-- :py:func:`galfitools.galout.getRads.getKappa`
-- :py:func:`galfitools.galout.getRads.getKappa2`
+- :py:func:`galfitools.galin.getSersic.getSersic`
+- :py:func:`galfitools.galin.initgal.InitGal`
 
 
 .. admonition:: Related GALFITools CLI commands
    :class: seealso
 
-   - :ref:`getKappa <routine-getKappa>`
-   - :ref:`getKappa2 <routine-getKappa2>`
+   - :ref:`getSersic <routine-getSersic>`
+   - :ref:`initgal <routine-initGal>`
 
 
 Example of a GALFIT file
@@ -100,18 +100,21 @@ parameter is free or fixed during fitting.
   10) 45.0     1             # Position angle (degrees)
    Z) 0                      # Skip this model in output image? (no=0)
 
+
+
 **Related GALFITools API**
 
-- :py:func:`galfitools.galout.getRads.getKappa`
-- :py:func:`galfitools.galout.getRads.getKappa2`
+- :py:func:`galfitools.galin.getSersic.getSersic`
+- :py:func:`galfitools.galin.initgal.InitGal`
+- :py:func:`galfitools.galin.checkGalFile.checkFile`
 
 
 .. admonition:: Related GALFITools CLI commands
    :class: seealso
 
-   - :ref:`getKappa <routine-getKappa>`
-   - :ref:`getKappa2 <routine-getKappa2>`
-
+   - :ref:`getSersic <routine-getSersic>`
+   - :ref:`initgal <routine-initGal>`
+   - :ref:`checkFile <routine-checkFile>`
 
 .. _concept-ds9:
 
@@ -123,15 +126,18 @@ and interactively examine astronomical data.
 
 **Related GALFITools API**
 
-- :py:func:`galfitools.galout.getRads.getKappa`
-- :py:func:`galfitools.galout.getRads.getKappa2`
-
+- :py:func:`galfitools.galin.MaskDs9.maskDs9`
+- :py:func:`galfitools.galout.PhotDs9.photDs9`
+- :py:func:`galfitools.sky.SkyDs9.SkyDs9`
 
 .. admonition:: Related GALFITools CLI commands
    :class: seealso
 
-   - :ref:`getKappa <routine-getKappa>`
-   - :ref:`getKappa2 <routine-getKappa2>`
+   - :ref:`maskDs9 <routine-maskDs9>`
+   - :ref:`photDs9 <routine-photDs9>`
+   - :ref:`skyDs9 <routine-skyDs9>`
+
+
 
 .. _concept-ds9-regions:
 
@@ -140,19 +146,24 @@ DS9 regions
 In DS9, a *region* is a user–defined geometric shape (circle, box,
 polygon, etc.) drawn on an image. Regions can mark sources, masks,
 or fitting boundaries, and can be saved to files that GALFITools 
-can read.
[…]
