+++ docs/index.rst
@@ -54,7 +54,7 @@ Indices and tables
 
 * :ref:`genindex`
 * :doc:`Overview <readme>`
-* :ref:`Concepts <concepts>`
+* :doc:`Concepts <concepts>`
 * :ref:`usage`
 * :ref:`api`
 * :ref:`modindex`
