+++ src/galfitools/shell/commands_mge.py
@@ -5,6 +5,73 @@ from galfitools.mge.SbProf import sbProf
 from galfitools.shell.prt import printWelcome
 
 
+# ---- helpers (small, testable) ----
+def _build_parser_mge() -> argparse.ArgumentParser:
+    p = argparse.ArgumentParser(
+        description="fits mge of Cappellari and formats to GALFIT"
+    )
+    p.add_argument("GalfitFile", help="GALFIT file to obtain the header options")
+    p.add_argument(
+        "Ds9regFile",
+        help=(
+            "the DS9 ellipse region file containing the galaxy. "
+            "Ignored if -xy option is used"
+        ),
+    )
+    p.add_argument(
+        "-t", "--twist", action="store_true", help="uses twist option for mge "
+    )
+    p.add_argument(
+        "-c",
+        "--center",
+        action="store_true",
+        help=(
+            "uses the center given in DS9 region file, "
+            "otherwise it will found the x,y peaks within DS9 ellipse"
+        ),
+    )
+    p.add_argument("-p", "--psf", type=float, help="the value of PSF sigma ", default=0)
+    p.add_argument(
+        "-gas",
+        "--gauss",
+        action="store_true",
+        help="uses gauss function for galfit file",
+    )
+    p.add_argument(
+        "-fser",
+        "--freeser",
+        action="store_true",
+        help="leaves the sersic index as a free parameter to fit",
+    )
+    p.add_argument(
+        "-fsk",
+        "--freesky",
+        action="store_true",
+        help="leaves the sky as a free parameter to fit",
+    )
+    p.add_argument(
+        "-ng",
+        "--numgauss",
+        type=int,
+        help="number of gaussians that will be used for galfit.",
+    )
+    p.add_argument(
+        "-xy",
+        "--xypos",
+        nargs=2,
+        type=int,
+        help="provides the (x y) position center of the object to fit",
+    )
+    p.add_argument("-e", "--ellip", type=float, help="ellipticity of object")
+    p.add_argument(
+        "-pa",
+        "--posang",
+        type=float,
+        help="position angle of object. Measured from Y-axis",
+    )
+    return p
+
+
 # check modify: remove initgauss
 def mainMGE():
     """
@@ -12,74 +79,9 @@ def mainMGE():
     This function serves as an example of an API.
     """
     printWelcome()
-    parser = argparse.ArgumentParser(
-        description="fits mge of Cappellari and formats to GALFIT"
-    )
+    parser = _build_parser_mge()
 
-    # parser.add_argument("image", help="the Mask image file to modify or create")
-    parser.add_argument("GalfitFile", help="GALFIT file to obtain the header options")
-    parser.add_argument(
-        "Ds9regFile",
-        help="the DS9 ellipse region file containing the galaxy."
-        + " Ignored if -xy option is used",
-    )
-    parser.add_argument(
-        "-t", "--twist", action="store_true", help="uses twist option for mge "
-    )
-    parser.add_argument(
-        "-c",
-        "--center",
-        action="store_true",
-        help="uses the center given in DS9 region file, "
-        + "otherwise it will found the x,y peaks within DS9 ellipse",
-    )
-    parser.add_argument(
-        "-p", "--psf", type=float, help="the value of PSF sigma ", default=0
-    )
-    parser.add_argument(
-        "-gas",
-        "--gauss",
-        action="store_true",
-        help="uses gauss function for galfit file",
-    )
-
-    parser.add_argument(
-        "-fser",
-        "--freeser",
-        action="store_true",
-        help="leaves the sersic index as a free parameter to fit",
-    )
[…]
