### 2025-08-23 b78fec5
command sky reformatted for tests
