+++ src/galfitools/shell/commands_sky.py
@@ -1,204 +1,130 @@
+# commands_sky.py  (rename to your actual module)
 import argparse
 
 from galfitools.shell.prt import printWelcome
 from galfitools.sky.GalfitSky import galfitSky
-
-# from galfitools.sky.Sky import sky
 from galfitools.sky.SkyDs9 import SkyDs9
 from galfitools.sky.SkyRing import SkyRing
 
 
-def mainGalfitSky():
-    """
-    Calls the galfitSky function based on argument parsing.
-    This function serves as an example of an API.
-    """
-
-    printWelcome()
-    parser = argparse.ArgumentParser(description="computes the sky using GALFIT")
-
-    parser.add_argument("image", help="the image file")
-    parser.add_argument("mask", help="the GALFIT mask file")
-
-    parser.add_argument(
+# ---------- parsers (small, testable) ----------
+def _build_parser_galfit_sky() -> argparse.ArgumentParser:
+    p = argparse.ArgumentParser(description="computes the sky using GALFIT")
+    p.add_argument("image", help="the image file")
+    p.add_argument("mask", help="the GALFIT mask file")
+    p.add_argument(
         "-s", "--scale", type=float, help="the plate scale. default = 1", default=1
     )
-
-    parser.add_argument(
+    p.add_argument(
         "-zp",
         "--mgzpt",
         type=float,
-        help="the magnitud zero point. default=25",
+        help="the magnitude zero point. default=25",
         default=25,
     )
-
-    parser.add_argument(
+    p.add_argument(
         "-x", "--xpos", type=float, help="the x position. default=1", default=1
     )
-    parser.add_argument(
+    p.add_argument(
         "-y", "--ypos", type=float, help="the y position. default=1", default=1
     )
-
-    parser.add_argument(
+    p.add_argument(
         "-is",
         "--initsky",
         type=float,
-        help="the initial sky value default=0",
+        help="the initial sky value. default=0",
         default=0,
     )
-
-    args = parser.parse_args()
-
-    imgname = args.image
-    maskfile = args.mask
-
-    mgzpt = args.mgzpt
-    scale = args.scale
-
-    X = args.xpos
-    Y = args.ypos
-
-    initsky = args.initsky
-
-    galfitSky(imgname, maskfile, mgzpt, scale, X, Y, initsky)
+    return p
 
 
-"""
-# Deprecated
-def mainSky():
-    ''' 
-    Calls the sky function based on argument parsing.
-    This function serves as an example of an API.
-    ''' 
-
-    printWelcome()
-
-    parser = argparse.ArgumentParser(
-        description="computes sky from a ds9 region box file"
+def _build_parser_sky_ds9() -> argparse.ArgumentParser:
+    p = argparse.ArgumentParser(
+        description="SkyDs9: computes sky background from a DS9 region file: Box, Ellipses, Polygons"
     )
-
-    parser.add_argument("image", help="the image file")
-    parser.add_argument("maskfile", help="the GALFIT Mask image file ")
-    parser.add_argument("Ds9regFile", help="the DS9 box region file")
-
-    args = parser.parse_args()
-
-    imgname = args.image
-    maskimage = args.maskfile
-    filereg = args.Ds9regFile
-
-    mean, sig = sky(imgname, maskimage, filereg)
-
-    print("Sky within 3 sigma:")
-
-    print("mean sky: {:.3f} ".format(mean))
-    print("std sky: {:.3f} ".format(sig))
-"""
-
-
-def mainSkyDs9():
-    """
-    Calls the SkyDs9 unction based on argument parsing.
-    This function serves as an example of an API.
-    """
-
-    printWelcome()
-
-    parser = argparse.ArgumentParser(
[…]
