+++ tests/test_maingalin.py
@@ -1,0 +1,257 @@
+import pytest
+
+import galfitools.shell.commands_galin as cli
+
+
+def test_mainGetStar(monkeypatch, capsys):
+    called = {}
+
+    def fake_getStar(image, regfile, imsize, center, sky, imout, sigma, sigout):
+        called.update(locals())
+
+    monkeypatch.setattr(cli, "getStar", fake_getStar)
+    monkeypatch.setattr(cli, "printWelcome", lambda: None)
+
+    argv = [
+        "img.fits",
+        "star.reg",
+        "64",
+        "--center",
+        "--sky",
+        "10.0",
+        "--out",
+        "cut.fits",
+        "--sigma",
+        "sig.fits",
+    ]
+    rc = cli.mainGetStar(argv)
+    assert rc == 0
+    assert called["image"] == "img.fits"
+    assert called["regfile"] == "star.reg"
+    assert called["imsize"] == 64
+    assert called["center"] is True
+    assert called["sky"] == 10.0
+    assert called["imout"] == "cut.fits"
+    assert called["sigma"] == "sig.fits"
+    out = capsys.readouterr().out
+    assert "Object fits file: cut.fits" in out
+
+
+def test_mainInitGal(monkeypatch):
+    captured = {}
+
+    def fake_InitGal(GalfitFile, number, p3, p4, p5, p6, p7, p8, p9, p10, numcomp):
+        captured.update(file=GalfitFile, number=number, p3=p3, numcomp=numcomp)
+
+    monkeypatch.setattr(cli, "InitGal", fake_InitGal)
+    monkeypatch.setattr(cli, "printWelcome", lambda: None)
+
+    argv = ["galfit.01", "-n", "3", "--param3", "10", "20", "--numcomp", "2"]
+    rc = cli.mainInitGal(argv)
+    assert rc == 0
+    assert captured["file"] == "galfit.01"
+    assert captured["number"] == 3
+    assert captured["p3"] == [10.0, 20.0]
+    assert captured["numcomp"] == 2
+
+
+def test_mainMaskDs9(monkeypatch):
+    called = {}
+
+    def fake_maskDs9(
+        MaskFile, RegFile, fill, image, border, borValue, skymean=None, skystd=None
+    ):
+        called.update(locals())
+
+    monkeypatch.setattr(cli, "maskDs9", fake_maskDs9)
+    monkeypatch.setattr(cli, "printWelcome", lambda: None)
+
+    argv = [
+        "mask.fits",
+        "regions.reg",
+        "--fill",
+        "5",
+        "--image",
+        "img.fits",
+        "--border",
+        "--borValue",
+        "2.0",
+        "--skymean",
+        "1.1",
+        "--skystd",
+        "0.9",
+    ]
+    rc = cli.mainMaskDs9(argv)
+    assert rc == 0
+    assert called["MaskFile"] == "mask.fits"
+    assert called["RegFile"] == "regions.reg"
+    assert called["fill"] == 5
+    assert called["image"] == "img.fits"
+    assert called["border"] is True
+    assert called["borValue"] == 2.0
+    assert called["skymean"] == 1.1
+    assert called["skystd"] == 0.9
+
+
+def test_mainMaskSky(monkeypatch, capsys):
+    got = {}
+
+    def fake_skyRem(image, mask, sky_mean, sky_sig, nsig, borValue, bor_flag):
+        got.update(locals())
+
+    def fake_maskDs9(*args, **kwargs):
+        got["maskDs9_called"] = True
+
+    monkeypatch.setattr(cli, "skyRem", fake_skyRem)
+    monkeypatch.setattr(cli, "maskDs9", fake_maskDs9)
[…]
