+++ src/galfitools/shell/commands_galin.py
@@ -1,7 +1,7 @@
 import argparse
 from collections import Counter
-
 import numpy as np
+
 from galfitools.galin.checkGalFile import checkFile
 from galfitools.galin.getBoxSizeDs9 import getBoxSizeDs9
 from galfitools.galin.getSersic import getSersic
@@ -16,18 +16,11 @@ from galfitools.galin.xy2fits import xy2fits
 from galfitools.shell.prt import printWelcome
 
 
-def mainGetStar():
-    """
-    Calls the getStar function based on argument parsing.
-    This function serves as an example of an API for getStar.
-    """
-
+def mainGetStar(argv=None) -> int:
     printWelcome()
-
     parser = argparse.ArgumentParser(
-        description="gets a image slice centered on the object peak"
+        description="gets an image slice centered on the object peak"
     )
-
     parser.add_argument("image", help="the image file to obtain the slice")
     parser.add_argument(
         "Ds9regFile", help="the DS9 ellipse region file containing the star"
@@ -37,8 +30,7 @@ def mainGetStar():
         "-c",
         "--center",
         action="store_true",
-        help="uses the center given in DS9 region file,"
-        + "otherwise it will find the x,y peaks within DS9 ellipse",
+        help="use the center in DS9 region; else find x,y peak within DS9 ellipse",
     )
     parser.add_argument(
         "-s", "--sky", type=float, help="the sky background to be removed"
@@ -46,9 +38,7 @@ def mainGetStar():
     parser.add_argument(
         "-o", "--out", type=str, help="the image output.", default="star.fits"
     )
-
     parser.add_argument("-sig", "--sigma", type=str, help="introduce the sigma image")
-
     parser.add_argument(
         "-so",
         "--sigout",
@@ -56,465 +46,281 @@ def mainGetStar():
         help="the sigma image output.",
         default="sigma.fits",
     )
+    args = parser.parse_args(argv)
 
-    args = parser.parse_args()
-
-    image = args.image
-    regfile = args.Ds9regFile
-    imsize = args.size
-    center = args.center
-
-    sky = args.sky
-    imout = args.out
-
-    sigma = args.sigma
-    sigout = args.sigout
-
-    getStar(image, regfile, imsize, center, sky, imout, sigma, sigout)
-
-    print("Done. Object fits file: {} created ".format(imout))
-
-    if sigma:
-        print("Done. sigma fits file: {} created ".format(sigout))
-
-
-def mainInitGal():
-    """
-    Calls the InitGal function based on argument parsing.
-    This function serves as an example of an API for InitGal.
-    """
-
-    printWelcome()
-
-
-    parser = argparse.ArgumentParser(
-        description="Creates GALFIT's input files with different initial parameters"
+    getStar(
+        args.image,
+        args.Ds9regFile,
+        args.size,
+        args.center,
+        args.sky,
+        args.out,
+        args.sigma,
+        args.sigout,
     )
+    print(f"Done. Object fits file: {args.out} created ")
+    if args.sigma:
+        print(f"Done. sigma fits file: {args.sigout} created ")
+    return 0
 
-    parser.add_argument("inFile", help="the galfit file galfit.XX ")
+
+def mainInitGal(argv=None) -> int:
+    printWelcome()
+    parser = argparse.ArgumentParser(
+        description="Creates GALFIT input files with different initial parameters"
+    )
+    parser.add_argument("inFile", help="the galfit file galfit.XX")
     parser.add_argument(
         "-n",
         "--number",
         type=int,
-        help="the number of files generated. Default = 1",
         default=1,
-    )
-
-    parser.add_argument(
-        "-p3",
-        "--param3",
-        nargs=2,
-        type=float,
-        help="range of values to give to the 3) model's parameter in format [min max] ",
[…]
