### 2025-08-23 0fa42a5
command galout reformatted for tests
