+++ src/galfitools/shell/commands_galout.py
@@ -1,5 +1,5 @@
+# commands_galout.py
 import argparse
-
 from galfitools.galin.galfit import Galfit
 from galfitools.galout.fitlog2csv import log2csv
 from galfitools.galout.getBarSize import getBarSize
@@ -24,1439 +24,425 @@ from galfitools.galout.showcube import displayCube
 from galfitools.shell.prt import printWelcome
 
 
-def mainPhotDs9():
-    """
-    Calls the photDs9 function based on argument parsing.
-    This function serves as an example of an API for photDs9.
-    """
-
+def mainPhotDs9(argv=None) -> int:
     printWelcome()
-
-    parser = argparse.ArgumentParser(
-        description="computes photometry from a Ds9 region "
-        + "file: Box, Ellipses and Polygons "
+    p = argparse.ArgumentParser(
+        description="computes photometry from a DS9 region file: Box, Ellipses and Polygons"
     )
-
-    parser.add_argument(
-        "ImageFile", help="the image file where the photometry will be computed"
-    )
-    parser.add_argument("RegFile", help="the DS9 region file")
-
-    parser.add_argument(
-        "-zp",
-        "--zeropoint",
-        type=float,
-        help="The value of the zero point. Default = 25 ",
-        default=25,
-    )
-
-    parser.add_argument("-m", "--mask", type=str, help="the mask file")
-
-    parser.add_argument(
-        "-sk",
-        "--sky",
-        default=0,
-        type=float,
-        help="Sky background value to be removed from image "
-        + "before photometry. Default = 0",
-    )
-
-    args = parser.parse_args()
-
-    ImageFile = args.ImageFile
-    RegFile = args.RegFile
-    maskfile = args.mask
-    zeropoint = args.zeropoint
-    sky = args.sky
-
-    mag, exptime = photDs9(ImageFile, RegFile, maskfile, zeropoint, sky)
-
-    line = "the exposition time is is: {} \n".format(exptime)
-    print(line)
-
-    line = "the magnitude from the ds9 region is: {:.2f} \n".format(mag)
-    print(line)
+    p.add_argument("ImageFile")
+    p.add_argument("RegFile")
+    p.add_argument("-zp", "--zeropoint", type=float, default=25)
+    p.add_argument("-m", "--mask", type=str)
+    p.add_argument("-sk", "--sky", type=float, default=0)
+    a = p.parse_args(argv)
+    mag, exptime = photDs9(a.ImageFile, a.RegFile, a.mask, a.zeropoint, a.sky)
+    print(f"the exposition time is is: {exptime} \n")
+    print(f"the magnitude from the ds9 region is: {mag:.2f} \n")
+    return 0
 
 
-# console scripts
-def mainGetBreak() -> None:
-    """
-    Calls the getBreak function based on argument parsing.
-    This function serves as an example of an API for getBreak.
-    """
-
+def mainGetBreak(argv=None) -> int:
     printWelcome()
-    # reading argument parsing
-
-    parser = argparse.ArgumentParser(
-        description="getBreak: gets the break radius from a set of Sersics "
+    p = argparse.ArgumentParser(
+        description="getBreak: gets the break radius from a set of Sersics"
     )
-
-    # required arguments
-    parser.add_argument(
-        "GalfitFile", help="Galfit File containing the Sersics or gaussians components"
-    )
-
-    parser.add_argument(
-        "-d", "--dis", type=int, help="Maximum distance among components", default=10
-    )
-
-    parser.add_argument(
-        "-n",
-        "--numcomp",
-        type=int,
-        help="Number of component where it'll obtain center of "
-        + "all components, default = 1 ",
[…]
