+++ src/galfitools/sky/GalfitSky.py
@@ -11,8 +11,6 @@ from galfitools.galin.MakeMask import ds9satbox
 from galfitools.mge.mge2galfit import PrintHeader
 from galfitools.mge.mge2galfit import PrintSky
 
-# change it and use skyRem instead
-
 
 def galfitSky(imgname, maskfile, mgzpt, scale, X, Y, sky) -> None:
     """Computes the sky using GALFIT
@@ -128,20 +126,3 @@ def galfitSky(imgname, maskfile, mgzpt, scale, X, Y, sky) -> None:
     fout1.close()
 
     print("To compute sky run: galfit sky.txt")
-
-
-####################################################
-
-# GALFIT functions
-
-#############################################################################
-#    End of program  ###################################
-#     ______________________________________________________________________
-#    /___/___/___/___/___/___/___/___/___/___/___/___/___/___/___/___/___/_/|
-#   |___|___|___|___|___|___|___|___|___|___|___|___|___|___|___|___|___|__/|
-#   |_|___|___|___|___|___|___|___|___|___|___|___|___|___|___|___|___|___|/|
-#   |___|___|___|___|___|___|___|___|___|___|___|___|___|___|___|___|___|__/|
-#   |_|___|___|___|___|___|___|___|___|___|___|___|___|___|___|___|___|___|/|
-#   |___|___|___|___|___|___|___|___|___|___|___|___|___|___|___|___|___|__/|
-#   |_|___|___|___|___|___|___|___|___|___|___|___|___|___|___|___|___|___|/
-##############################################################################
