+++ tests/test_maingalout.py
@@ -1,4 +1,5 @@
 import pytest
+import re
 import galfitools.shell.commands_galout as cli
 
 
@@ -48,6 +49,51 @@ def test_mainGetBreak2(monkeypatch, capsys):
     assert "The break radius is 12.50" in out
 
 
+def test_mainFWHM(monkeypatch, capsys):
+    # Stub getFWHM so we don’t need real galfit files
+    monkeypatch.setattr(cli, "getFWHM", lambda *a, **k: (8.5, 3, 42.0))
+
+    # Fake Galfit.ReadHead().scale
+    class Head:
+        scale = 0.5
+
+    monkeypatch.setattr(
+        cli, "Galfit", lambda f: type("G", (), {"ReadHead": lambda self: Head()})()
+    )
+
+    # Silence welcome banner
+    monkeypatch.setattr(cli, "printWelcome", lambda: None)
+
+    # Call with minimal arguments
+    rc = cli.mainFWHM(["galfit.01", "-d", "5", "-n", "2", "-pa", "37.5"])
+    assert rc == 0
+
+    out = capsys.readouterr().out
+    # Validate key parts of output
+    assert "number of model components:  3" in out
+    assert "theta value of : 42.00" in out
+    assert "FWHM is 8.50 pixels or 4.25" in out  # 8.5 * 0.5 = 4.25
+
+
+def test_mainKappa(monkeypatch, capsys):
+    monkeypatch.setattr(cli, "getKappa", lambda *a: (7.0, 4, 30.0))
+
+    class Head:
+        scale = 0.5
+
+    monkeypatch.setattr(
+        cli, "Galfit", lambda f: type("G", (), {"ReadHead": lambda self: Head()})()
+    )
+    monkeypatch.setattr(cli, "printWelcome", lambda: None)
+
+    rc = cli.mainKappa(["galfit.01"])
+    assert rc == 0
+
+    out = capsys.readouterr().out
+    norm = " ".join(out.split()).lower()  # collapse spaces, lower-case
+    assert 'kappa radius is 7.00 pixels or 3.50 "' in norm
+
+
 def test_mainKappa2(monkeypatch, capsys):
     monkeypatch.setattr(cli, "getKappa2", lambda *a: (7.0, 4, 30.0))
 
@@ -58,10 +104,85 @@ def test_mainKappa2(monkeypatch, capsys):
         cli, "Galfit", lambda f: type("G", (), {"ReadHead": lambda self: Head()})()
     )
     monkeypatch.setattr(cli, "printWelcome", lambda: None)
+
     rc = cli.mainKappa2(["galfit.01"])
     assert rc == 0
+
     out = capsys.readouterr().out
-    assert 'kappa radius is 7.00 pixels or 3.50 "' in out
+    norm = " ".join(out.split()).lower()
+    assert 'kappa radius is 7.00 pixels or 3.50 "' in norm
+
+
+def test_mainGetReComp_default_half(monkeypatch, capsys):
+    """
+    fracrad defaults to 0.5 → the 'Mean Surface Brightness at effective radius' line should print.
+    """
+    # Stub heavy computation
+    # Returns: EffRad, totmag, meanme, me, N, theta
+    monkeypatch.setattr(
+        cli, "getReComp", lambda *a, **k: (12.0, 19.25, 21.10, 20.50, 3, 37.0)
+    )
+
+    # Fake Galfit.ReadHead().scale
+    class Head:
+        scale = 0.4  # arcsec / pixel
+
+    monkeypatch.setattr(
+        cli, "Galfit", lambda f: type("G", (), {"ReadHead": lambda self: Head()})()
+    )
+
+    # Silence banner
[…]
