+++ src/galfitools/sky/SkyRing.py
@@ -60,7 +60,7 @@ def SkyRing(image, mask, ds9regfile, width, center, outliers):
 
     assert os.path.exists(image), "image file does not exists"
 
-    if mask:
+    if mask:  # pragma: no cover
         assert os.path.exists(mask), "mask file does not exists"
 
     # end input
@@ -76,12 +76,12 @@ def SkyRing(image, mask, ds9regfile, width, center, outliers):
         hdumask = fits.open(mask)
         maskimg = hdumask[0].data
         hdumask.close()
-    else:
+    else:  # pragma: no cover
         maskimg = np.array([])
 
     (ncol, nrow) = GetAxis(image)
 
-    if center:
+    if center:  # pragma: no cover
         print("center of ds9 ellipse region will be used")
         xpeak, ypeak = xpos, ypos
     else:
@@ -184,7 +184,7 @@ class SkyCal:
 
         if MaskImg.any():
             self.maskimg = MaskImg.copy()
-        else:
+        else:  # pragma: no cover
             self.maskimg = np.array([])
 
         # hdu = fits.open(ImageFile)
@@ -200,8 +200,8 @@ class SkyCal:
 
         self.R = Rkron
 
-        if self.Rinit > Rkron:  # avoid radius greater than the border
-
+        if self.Rinit > Rkron:  # pragma: no cover
+            # avoid radius greater than the border
             self.Rinit = Rkron / 3  # is this number ok?
             print("Rinit is greater than image size")
             print("using Rinit = {:0.2f} ".format(self.Rinit))
@@ -219,7 +219,7 @@ class SkyCal:
             ]  # logical patch mask image
 
             self.invpatch = np.logical_not(patch)
-        else:
+        else:  # pragma: no cover
             self.invpatch = np.array([])
 
         Rings = np.arange(self.Rinit, self.R, self.width)  # anillos de tamaño width
@@ -275,7 +275,8 @@ class SkyCal:
 
         ########################################
         ########################################
-        if self.outliers:  # eliminate top 80% and bottom 20%
+        if self.outliers:  # pragma: no cover
+            # eliminate top 80% and bottom 20%
             print("removing top 80% and bottom 20% for every ring")
 
         # computing sky in every ring
@@ -295,7 +296,7 @@ class SkyCal:
 
             if self.outliers:  # eliminate top 80% and bottom 20%
                 imgpatch = flatimg[bot:top]
-            else:
+            else:  # pragma: no cover
                 imgpatch = flatimg
 
             mean = np.mean(imgpatch)
@@ -346,7 +347,7 @@ class SkyCal:
             count += 1
             idx += 1
 
-            if idx == (len(Rings) - 1):
+            if idx == (len(Rings) - 1):  # pragma: no cover
                 print("The edge of image has been reached. Sky can not be computed")
                 return 0, 0, 0, 0
 
@@ -372,14 +373,14 @@ class SkyCal:
 
         maskring = masksky == ring
 
-        if self.invpatch.any():
+        if self.invpatch.any():  # pragma: no cover
             maskring = maskring * self.invpatch
 
         ringcont = 0
 
         while not (maskring.any()) and (ringcont < 10):
 
-            if ringcont == 0:
+            if ringcont == 0:  # pragma: no cover
                 print("Selecting next ring ")
 
             idx += 1
@@ -392,7 +393,7 @@ class SkyCal:
 
             ringcont += 1  # avoid eternal loop
 
-        if ringcont == 10:
+        if ringcont == 10:  # pragma: no cover
             print("max. iteration reached. I couldn't find a ring")
[…]
