+++ src/galfitools/mge/mge2galfit.py
@@ -127,13 +127,13 @@ def mge2gal(
     if (maskfile == "None") or (maskfile == "none"):
         maskfile = None
 
-    if (sigfile == "None") or (sigfile == "none"):
+    if (sigfile == "None") or (sigfile == "none"):  # pragma: no cover
         sigfile = None
 
-    if (psfile == "None") or (psfile == "none"):
+    if (psfile == "None") or (psfile == "none"):  # pragma: no cover
         psfile = None
 
-    if regu:
+    if regu:  # pragma: no cover
         try:
             from mgefit.mge_fit_sectors_twist_regularized import (
                 mge_fit_sectors_twist_regularized,
@@ -147,7 +147,7 @@ def mge2gal(
     ##########################################
     ############################################
 
-    if regu:
+    if regu:  # pragma: no cover
         print("regularization mode activated")
 
     root_ext = os.path.splitext(image)
@@ -209,7 +209,7 @@ def mge2gal(
     eps = 0
     theta = 0
 
-    if regfile:
+    if regfile:  # pragma: no cover
         if os.path.isfile(regfile):
 
             obj, xpos, ypos, rx, ry, angle = GetInfoEllip(regfile)
@@ -248,7 +248,7 @@ def mge2gal(
 
     # plt.clf()
 
-    if twist:
+    if twist:  # pragma: no cover
         print("twist mode mge fit sectors is used")
         # Perform galaxy photometry
         if maskfile:
@@ -348,7 +348,7 @@ def mge2gal(
         tolpsf = 0.001
         if np.abs(psf) > tolpsf:
 
-            if regu:
+            if regu:  # pragma: no cover
 
                 m = mge_fit_sectors_regularized(
                     s.radius,
@@ -379,7 +379,7 @@ def mge2gal(
         else:
             print("No convolution")
 
-            if regu:
+            if regu:  # pragma: no cover
                 m = mge_fit_sectors_regularized(
                     s.radius,
                     s.angle,
@@ -408,7 +408,7 @@ def mge2gal(
 
         plt.savefig(namepng)
 
-    if twist:
+    if twist:  # pragma: no cover
 
         (counts, sigma, axisrat, pa) = m.sol
 
@@ -426,7 +426,7 @@ def mge2gal(
     #  print GALFIT files
     #####################
 
-    if gauss:
+    if gauss:  # pragma: no cover
         parfile = "mgeGALFIT.txt"
     else:
         parfile = "mseGALFIT.txt"
@@ -434,7 +434,7 @@ def mge2gal(
     outname = TNAM
 
     rmsname = sigfile
-    if psfile:
+    if psfile:  # pragma: no cover
         psfname = psfile
     else:
         psfname = "None"
@@ -464,7 +464,7 @@ def mge2gal(
 
     K = gammaincinv(1, 0.5)
 
-    if psfile:
+    if psfile:  # pragma: no cover
         (ncol, nrow) = GetAxis(psfname)
         convbox = ncol + 1
         convboxy = nrow + 1
@@ -492,7 +492,7 @@ def mge2gal(
         0,
     )
 
-    if freeser:
+    if freeser:  # pragma: no cover
         serfit = 1
     else:
         serfit = 0
@@ -500,12 +500,12 @@ def mge2gal(
     # removing the last components:
     totGauss = len(counts)
 
-    if not (numgauss):
+    if not (numgauss):  # pragma: no cover
         numgauss = totGauss
 
     print("total number of gaussians by mge_fit_sectors: ", totGauss)
 
-    if numgauss > totGauss:
+    if numgauss > totGauss:  # pragma: no cover
         print("number of gaussians is greater than the ones fitted by mge_fit_sectors")
[…]
