+++ tests/test_galin.py
@@ -8,6 +8,10 @@ import numpy as np
 from collections import Counter
 from astropy.io import fits
 
+import stat
+import re
+from pathlib import Path
+
 from galfitools.galin.getStar import getStar
 from galfitools.galin.initgal import InitGal
 from galfitools.galin.MaskDs9 import maskDs9
@@ -106,6 +110,175 @@ def test_InitGal():
         os.remove(namefile)
 
     return None
+
+
+def _write_minimal_galfit(path: Path):
+    """
+    Create a minimal GALFIT input file with:
+      - component 1: lines 0), 3) .. 10)
+      - component 2: lines 0), 3) .. 10)
+      - sky component: 0) sky and a 3) line (must NOT be changed)
+    """
+    lines = []
+    # component 1
+    lines += [
+        "0) sersic",
+        "3) 1.00",  # typical radius-like parameter
+        "4) 2.00",
+        "5) 3.00",
+        "6) 4.00",
+        "7) 5.00",
+        "8) 6.00",
+        "9) 7.00",
+        "10) 8.00",
+    ]
+    # component 2
+    lines += [
+        "0) sersic",
+        "3) 10.00",
+        "4) 20.00",
+        "5) 30.00",
+        "6) 40.00",
+        "7) 50.00",
+        "8) 60.00",
+        "9) 70.00",
+        "10) 80.00",
+    ]
+    # sky component at end (flagsky3 must prevent changing its 3))
+    lines += [
+        "0) sky",
+        "3) 999.00",
+    ]
+    path.write_text("\n".join(lines) + "\n", encoding="utf-8")
+
+
+def _extract_param_values(text: str, token: str):
+    """
+    Return a list of floats for a given parameter token like '3)' found at
+    the beginning of a line (possibly with a leading space that InitGal adds).
+    """
+    vals = []
+    for line in text.splitlines():
+        m = re.match(rf"^\s*{re.escape(token)}\)\s+([+-]?\d+(\.\d+)?)", line)
+        if m:
+            vals.append(float(m.group(1)))
+    return vals
+
+
+def test_initgal_generates_files_and_bash(tmp_path):
+    gal_in = tmp_path / "galfit.gal"
+    _write_minimal_galfit(gal_in)
+
+    # Change cwd so InitGal.MakeBash() sees tmp_path/*.gal
+    old_cwd = os.getcwd()
+    os.chdir(tmp_path)
+    try:
+        InitGal(
+            str(gal_in),
+            number=2,
+            param3=(0.5, 1.5),
+            param4=(1.5, 2.5),
+            param5=(2.5, 3.5),
+            param6=(3.5, 4.5),
+            param7=(4.5, 5.5),
+            param8=(5.5, 6.5),
+            param9=(6.5, 7.5),
+            param10=(7.5, 8.5),
+            numcomp=0,
+        )
+
+        # Now check generated files
+        files = sorted(p.name for p in tmp_path.glob("*.gal"))
+        assert "galfit-1.gal" in files
+        assert "galfit-2.gal" in files
+
+        # Bash script created and executable
+        bash = tmp_path / "rungalfit.sh"
+        assert bash.exists()
+        btxt = bash.read_text()
+        assert "galfit galfit-1.gal" in btxt
+        assert "galfit galfit-2.gal" in btxt
+    finally:
+        os.chdir(old_cwd)
+
+
+def test_initgal_only_target_component_changes(tmp_path: Path):
+    """numcomp=2 → only component 2 parameters are randomized; component 1 stays at original values; sky 3) unchanged."""
+    gal_in = tmp_path / "galfit.gal"
+    _write_minimal_galfit(gal_in)
+
[…]
