+++ src/galfitools/galout/getRads.py
@@ -124,14 +124,14 @@ def getBreak(
 
     maskgal = comps.Active == 1
 
-    if angle:
+    if angle:  # pragma: no cover
         theta = angle
     else:
         theta = comps.PosAng[maskgal][-1]
 
     N = numComps(comps, "all")
 
-    if N == 0:
+    if N == 0:  # pragma: no cover
         print("not enough number of components to compute Re")
         print("exiting..")
         sys.exit(1)
@@ -140,7 +140,7 @@ def getBreak(
     #  computing the slope
     #########################
 
-    if plot:
+    if plot:  # pragma: no cover
 
         if ranx:
             (xmin, xmax) = ranx[0], ranx[1]
@@ -161,13 +161,13 @@ def getBreak(
         plt.minorticks_on()
         plt.savefig("Break.png")
 
-    if quick:
+    if quick:  # pragma: no cover
 
         rbreak = GetBreak().FindBreak(comps, theta, inicomp)
 
     else:
 
-        if random:
+        if random:  # pragma: no cover
 
             radius = np.random.random(random) * comps.Rad[maskgal][inicomp]
 
@@ -244,7 +244,9 @@ class GetBreak:
 
     """
 
-    def GalBreak(self, R: list, comps: GalComps, theta: float) -> float:
+    def GalBreak(
+        self, R: list, comps: GalComps, theta: float
+    ) -> float:  # pragma: no cover
 
         maskgal = comps.Active == 1  # using active components only
 
@@ -286,7 +288,9 @@ class GetBreak:
 
     """
 
-    def FindBreak(self, comps: GalComps, theta: float, initial_comp: int) -> float:
+    def FindBreak(
+        self, comps: GalComps, theta: float, initial_comp: int
+    ) -> float:  # pragma: no cover
 
         maskgal = comps.Active == 1  # using active components only
 
@@ -445,7 +449,7 @@ class GetBreak:
 
     def SlopeSer(
         self, R: float, Ie: list, Re: list, n: list, q: list, pa: list, theta: float
-    ) -> float:
+    ) -> float:  # pragma: no cover
         """slope from sersic function to a determined R"""
 
         Rcor = GetRadAng(R, q, pa, theta)
@@ -549,7 +553,7 @@ def getBreak2(
 
     N = numComps(comps, "all")
 
-    if N == 0:
+    if N == 0:  # pragma: no cover
         print("not enough number of components to compute break radius")
         print("exiting..")
         sys.exit(1)
@@ -568,7 +572,7 @@ def getBreak2(
     lrad = np.log10(R)
     slp = GetSlope().GalSlope(R, comps, theta)
 
-    if plot:
+    if plot:  # pragma: no cover
         plt.close()
         plt.plot(R, slp)
 
@@ -585,7 +589,7 @@ def getBreak2(
     # yspl2d = yspl.derivative(n=2)
     yspl1d = yspl.derivative(n=1)
 
-    if plot:
+    if plot:  # pragma: no cover
         plt.close()
         plt.plot(10 ** (lrad), yspl1d(lrad))
         plt.xlabel("Rad (pixels)")
@@ -692,7 +696,7 @@ def getKappa2(
 
     N = numComps(comps, "all")
 
-    if N == 0:
+    if N == 0:  # pragma: no cover
         print("not enough number of components to compute kappa radius")
         print("exiting..")
         sys.exit(1)
@@ -719,7 +723,7 @@ def getKappa2(
 
     kap1d = np.abs(yspl1d(lrad)) / (1 + slp**2) ** (3 / 2)
 
-    if plot:
+    if plot:  # pragma: no cover
         plt.close()
         plt.plot(10 ** (lrad), kap1d)
         plt.xlabel("Rad (pixels)")
[…]
