### 2025-08-28 1b95d1c
skyDs9 now computes surface brightness of sky
