### 2025-08-29 707f7d1
skyRing now computes surface brightness of sky
