### 2025-08-30 54393a7
msg  outputs of position angle in galouts commands changed
