+++ src/galfitools/shell/commands_galout.py
@@ -68,7 +68,11 @@ def mainGetBreak(argv=None) -> int:
         a.ranx,
     )
     print("number of model components: ", N)
-    print(f"Using a theta value of: {theta:.2f} degrees\n")
+    if a.angle:
+        print(f"position angle: {theta:.2f} degrees\n")
+    else:
+        print(f"position angle: {theta:.2f} degrees, taken from component {N}\n")
+
     plate = Galfit(a.GalfitFile).ReadHead().scale
     print(f'The break radius is {rbreak:.2f} pixels or {rbreak*plate:.2f} "  \n')
     return 0
@@ -91,7 +95,10 @@ def mainGetBreak2(argv=None) -> int:
     )
     plate = Galfit(a.GalfitFile).ReadHead().scale
     print("number of model components: ", N)
-    print(f"Using a theta value of: {theta:.2f} degrees\n")
+    if a.angle:
+        print(f"position angle: {theta:.2f} degrees\n")
+    else:
+        print(f"position angle: {theta:.2f} degrees, taken from component {N}\n")
     print(f'The break radius is {rbreak:.2f} pixels or {rbreak*plate:.2f} " \n')
     return 0
 
@@ -111,7 +118,7 @@ def mainGetBarSize(argv=None) -> int:
     rbar, N, theta = getBarSize(a.GalfitFile, a.dis, a.numcomp, a.plot, a.ranx, a.out)
     plate = Galfit(a.GalfitFile).ReadHead().scale
     print("number of model components: ", N)
-    print(f"Using a theta value of: {theta:.2f} degrees\n")
+
     print(f'The bar size is {rbar:.2f} pixels or {rbar*plate:.2f} " \n')
     return 0
 
@@ -129,7 +136,10 @@ def mainFWHM(argv=None) -> int:
     fwhm, N, theta = getFWHM(a.GalfitFile, a.dis, a.angle, a.numcomp)
     plate = Galfit(a.GalfitFile).ReadHead().scale
     print("number of model components: ", N)
-    print(f"Using a theta value of : {theta:.2f} degrees\n")
+    if a.angle:
+        print(f"position angle: {theta:.2f} degrees\n")
+    else:
+        print(f"position angle: {theta:.2f} degrees, taken from component {N}\n")
     print(f'The FWHM is {fwhm:.2f} pixels or {fwhm*plate:.2f} " \n')
     return 0
 
@@ -162,6 +172,11 @@ def mainKappa(argv=None) -> int:
     )
     plate = Galfit(a.GalfitFile).ReadHead().scale
     print("number of model components: ", N)
+    if a.angle:
+        print(f"position angle: {theta:.2f} degrees\n")
+    else:
+        print(f"position angle: {theta:.2f} degrees, taken from component {N}\n")
+
     print(f'The Kappa radius  is {rkappa:.2f} pixels or {rkappa*plate:.2f} " \n')
     return 0
 
@@ -183,7 +198,11 @@ def mainKappa2(argv=None) -> int:
     )
     plate = Galfit(a.GalfitFile).ReadHead().scale
     print("number of model components: ", N)
-    print(f"Using a theta value of: {theta:.2f} degrees\n")
+    if a.angle:
+        print(f"position angle: {theta:.2f} degrees\n")
+    else:
+        print(f"position angle: {theta:.2f} degrees, taken from component {N}\n")
+
     print(f'The kappa radius is {rkappa:.2f} pixels or {rkappa*plate:.2f} " \n')
     return 0
 
@@ -204,7 +223,11 @@ def mainGetReComp(argv=None) -> int:
     )
     plate = Galfit(a.GalfitFile).ReadHead().scale
     print("number of model components: ", N)
[…]
