### 2025-08-30 411e52a
msg  outputs of position angle in galouts commands changed
