### 2025-09-01 cead1e0
documentation and paper updated
