### 2025-09-04 #220 @canorve
Dear reviewer @Jammy2211 :

> My main critique is that the package assumes the user has a lot of Astronomy experience and knowledge in terms of how its documented and phrased ...

We have added a new Concepts section in the documentation, which includes a primer on key concepts, GALFIT example file, inline definitions of terms, and a brief Audience statement on the main page. We also reduced the use of jargon and added cross-links to definitions. For each concept, we now provide the related API functions and CLI commands directly below, making it easier for users without a deep astronomy background to navigate the package.

>     I can’t see the statement of need in the docs, only the JOSS paper has one, maybe move it to here: https://galfitools.readthedocs.io/en/latest/index.html

I have added a clear Statement of Need to the index page of the documentation, adapted from the JOSS paper, so that new users can immediately understand the motivation for GALFITools.

>     Please include the supported Python versions (e..g. 3.10 - 3.12 ?) in the installation instructions.

The supported Python versions (3.10–3.13) have been added to the installation instructions.

>     The link to “install GALFIT” on the installation page goes to the FAQ, which has no installation instructions. Link here instead: https://users.obs.carnegiescience.edu/peng/work/galfit/galfit.html

I have updated the installation page so that the link now points to the installation file added to the project’s GitHub repository. The guide provides step-by-step instructions and replaces the previous FAQ link.

>      Instructions on needing to install pytest-cov for unit testing.

I have updated the contributing/testing instructions (`CONTRIBUTING.rst`)  to specify that pytest-cov is required.
