### 2025-09-04 #220 @canorve
Dear reviewer @tepickering :

>     I found the statement of need to be sufficient because the package has a very specific target audience: the users of galfit. As such, some prior knowledge of galfit and its capabilities can be assumed. A significant background in astronomy and astronomical image analysis is a prerequisite for effectively using galfit and interpreting its results. The same prerequisite applies to this package.

Thank you. The statement of need has also been included in the documentation, with slight modifications for clarification.

 > 	The link to the datafile used for the example, galfit.01, points to a file within the repository tree. As it's currently linked, it is a multi-step process to download into a usable format. One option would be to use a link to the raw file, e.g. https://raw.githubusercontent.com/canorve/GALFITools/refs/heads/master/docs/galfit.01, which would facilitate right-click->"save as", or include the file within the documentation tree and serve from RTD.

The link to galfit.01 has been updated so that it now points directly to the raw file, allowing immediate download.

>     The links to the "Module Reference", "Usage guide", and "API Reference" are kinda buried at the bottom of the documentation index page. It would be nice to feature them more prominently so they're easier to find, e.g. in the "Documentation" section which currently only links recursively back to the main page.

I have moved the "Module Reference" and "Usage Guide" to a more prominent location in the documentation index to make them easier to find.

>     Several requested badges are missing.

I have added the remaining requested badges to the README.

>     I tested locally under python 3.13 and everything passed. I would recommend adding 3.13 to the CI test matrix.

Python 3.13 has been added to the CI test matrix.
