### 2025-09-21 #220 @hamogu
@tepickering and @Jammy2211: Thanks for your thorough review and good suggestions and to @canorve for addresses them; It hope the team found it all useful!

@tepickering and @Jammy2211: Can you take a look at the changes to the package and see if they address the points you raised in the review? If so, please update the "Final approval (post-review)" in your review and let me know here (I only get a notification from GH if there is a new comment, not if an old comment is updated); if not we welcome your feedback what else can be improved!
