### 2025-10-30 #220 @hamogu
To package authors and reviewers: We are happy to invite you to the PyOpenSci Slack (at least for now, we are probably migrating to a different platform, e.g. Zulip soon). There is a lot of discussion about best practices for packaging and  very helpful community to help you out if you ever get stuck.
If you are interested, just contact me with the email address you'd like to me send the invite to.
