### 2025-10-30 #220 @hamogu
@canorve Please let me know when you've submitted this to JOSS (and remember to link to this issue, so JOSS knows that they don't have to do a code review again!) so that I can follow along and update the status here, including the JOSS DOI when it's accepted by JOSS, too.
