### 2025-11-05 #220 @canorve
> [@canorve](https://github.com/canorve) When you opened the issue, you ticked "submit to JOSS", but not "The package contains a paper.md matching [JOSS's requirements](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) with a high-level description in the package root or in inst/."
> Of course, the package DOES have a `paper.md` file and our reviewers thinks it's a good one. Just checking before we pass that on to JOSS: Is there anything that you believe still needs to be done to the `paper.md` or can we just go back and set the tick for "The package contains a paper.md matching ..."?

Thank you for checking. The paper.md file is complete and matches JOSS’s requirements. There are no further changes needed, so please go ahead and mark the item as fulfilled.
