### 2025-11-05 6bf34ef
pyopensci badge added to readme
