### 2025-11-06 #220 @canorve
Thank you to the reviewers and the editor for your valuable feedback on the GALFITools package.

Dear editor @hamogu , I have added the badge and completed the post-review survey. I just wanted to let you know.
