### 2025-11-20 13ec806
workflow actions for pdf generator
