+++ .github/workflows/draft-pdf.yml
@@ -19,7 +19,7 @@ jobs:
         with:
           journal: joss
           # This should be the path to the paper within your repo.
-          paper-path: paper.md
+          paper-path: paper/paper.md
       - name: Upload
         uses: actions/upload-artifact@v4
         with:
@@ -27,7 +27,7 @@ jobs:
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
-          path: paper.pdf
+          path: paper/paper.pdf
 
 
       - name: Commit PDF to repository
@@ -35,7 +35,7 @@ jobs:
         with:
           message: '(auto) Paper PDF Draft'
           # This should be the path to the paper within your repo.
-          add: 'paper.pdf' # 'paper/*.pdf' to commit all PDFs in the paper directory
+          add: 'paper/paper.pdf' # 'paper/*.pdf' to commit all PDFs in the paper directory
 
 
 
