### 2025-11-22 #220 @canorve
> [@canorve](https://github.com/canorve) Please let me know when you've submitted this to JOSS (and remember to link to this issue, so JOSS knows that they don't have to do a code review again!) so that I can follow along and update the status here, including the JOSS DOI when it's accepted by JOSS, too.

Hi @hamogu,  I'm just letting you know that I've submitted this package to JOSS
