### 2025-11-30 #220 @hamogu
@canorve Do you happen to know what the issue number in the JOSS reviews is (or has it not been opened yet maybe because of the Thanksgiving holiday in the USA)? I've looked several times, but can't find it in the JOSS queue yet.

I'm asking so that I can subscribe to that issue and (a) make sure JOSS doesn't accidentally kick-off another review process (they don't need to since we already did that) and (b) that I can make sure our metadata is updated with the JOSS DOI when they accept it so that the journal reference gets properly cross-linked from the PyOpenSci listing. I'll keep checking.
