### 2025-11-30 #220 @canorve
Thanks for checking. I reached out to JOSS, and they confirmed that the submission is in their system, but the review issue hasn’t been opened yet. As soon as the issue is opened, I will let you know

Message posted in JOSS: This paper is submitted but the review hasn't started.
