### 2025-12-08 b0fa6a9
Initial editorial changes for publication in JOSS
