+++ paper/paper.bib
@@ -6,7 +6,7 @@
 
 ```bibtex
 @article{peng02,
-  author       = {Peng, C. Y. and Ho, L. C. and Impey, C. D. and Rix, H.-W.},
+  author       = {{Peng}, Chien Y. and Ho, L. C. and Impey, C. D. and Rix, H.-W.},
   title        = {Detailed Structural Decomposition of Galaxy Images},
   journal      = {The Astronomical Journal},
   volume       = {124},
@@ -28,8 +28,8 @@
 
 @ARTICLE{peng10,
        author = {{Peng}, Chien Y. and {Ho}, Luis C. and {Impey}, Chris D. and {Rix}, Hans-Walter},
-        title = "{Detailed Decomposition of Galaxy Images. II. Beyond Axisymmetric Models}",
-      journal = {\aj},
+        title = {Detailed Decomposition of Galaxy Images. II. Beyond Axisymmetric Models},
+      journal = {The Astronomical Journal},
      keywords = {galaxies: bulges, galaxies: fundamental parameters, galaxies: structure, techniques: image processing, techniques: photometric, Astrophysics - Cosmology and Nongalactic Astrophysics},
          year = 2010,
         month = jun,
@@ -71,7 +71,7 @@ volume = {799},
 number = {2},
 pages = {226},
 author = {Erwin, Peter},
-title = {IMFIT: A FAST, FLEXIBLE NEW PROGRAM FOR ASTRONOMICAL IMAGE FITTING},
+title = {Imfit: a fast, flexible new program for astronomical image fitting},
 journal = {The Astrophysical Journal},
 abstract = {I describe a new, open-source astronomical image-fitting program called imfit, specialized for galaxies but potentially useful for other sources, which is fast, flexible, and highly extensible. A key characteristic of the program is an object-oriented design that allows new types of image components (two-dimensional surface-brightness functions) to be easily written and added to the program. Image functions provided with imfit include the usual suspects for galaxy decompositions (Sérsic, exponential, Gaussian), along with Core-Sérsic and broken-exponential profiles, elliptical rings, and three components that perform line-of-sight integration through three-dimensional luminosity–density models of disks and rings seen at arbitrary inclinations. Available minimization algorithms include Levenberg–Marquardt, Nelder–Mead simplex, and Differential Evolution, allowing trade-offs between speed and decreased sensitivity to local minima in the fit landscape. Minimization can be done using the standard χ2 statistic (using either data or model values to estimate per-pixel Gaussian errors, or else user-supplied error images) or Poisson-based maximum-likelihood statistics; the latter approach is particularly appropriate for cases of Poisson data in the low-count regime. I show that fitting low-signal-to-noise ratio galaxy images using χ2 minimization and individual-pixel Gaussian uncertainties can lead to significant biases in fitted parameter values, which are avoided if a Poisson-based statistic is used; this is true even when Gaussian read noise is present.}
 }
