### 2025-12-30 b32ffcf
small edits for joss
