### 2025-12-31 #220 @hamogu
Congratulations on the JOSS acceptance! 🥂
I've updated the JOSS DOI in the metadata and will close this issue, too! The package listing on the website should update soon through a scripts that's run one a day or so.
